==> src/02_text_trends/00_antiquated/03_diversity-h3.Rmd <==
---
title: "Hypothesis 3 (Diversity Sample)"
description: "The use of terms like genetic ancestry as well as national and continental terms have outpaced racial and ethnic terms in biomedical literature."
output: html_document
---

To test Hypothesis 3, we parsed our dictionary of 2,200+ population terms into a set of more nuanced analytic categories. Our main objective is to test Fujimura and Rajagopalan's (2011) assertion that researchers now use biogeographic terms more often than race and ethnicity in their studies. Once again, following Panofsky and Bliss (2017), we looked at the use of continent, country, and directional terminology, US-specific terms (corresponding to the US Census and OMB Directive 15 labeling schema), a category of general population terms (i.e. ancestry, race, ethnicity, population, minority, etc.), and a category of all available ethnic terms in our compiled dictionary (see Appendix below). 

```{r prereqs, message = FALSE, results = FALSE, echo=FALSE}
rm(list = ls())

for (pkg in c("tidyverse", "igraph", "tidytext", "ggplot2", "DT",
              "grid", "gridExtra", "reshape2", "extrafont",
               "plotly", "stringi", "networkD3")) {library(pkg, character.only = TRUE)}

setwd("~/Documents/Diversity/Data")
# loading the .csv file 
text_data <- read_csv("diversity_corpus.csv") %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018") 
# tokenizing the abstract data into words 
abstract_data <- text_data %>% 
  unnest_tokens(word, abstract) %>% 
  anti_join(stop_words)
# adding custom set of stopwords 
my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier"))
abstract_data <- abstract_data %>% anti_join(my_stopwords); rm(my_stopwords)

# lets draw all of our strings from the diversity_dictionary
setwd("~/Documents/Diversity/Data")
divictionary <- read_csv("diversity_project - h1_categories.csv") 
# pulling our list of population
all_pop_terms <- read_csv("diversity_project - h2h3_categories.csv") 
all_pop_terms <- paste(c("\\b(?i)(zcx", all_pop_terms$term, "zxc)\\b"), collapse = "|")
# pulling in concatenated population terms from our dataset 
us_specific_terms <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(category == "us_specific")
general_pop_terms <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(category == "general_terms")
continental_terms <- read_csv("diversity_project - h2h3_categories.csv") %>%
  filter(category == "continental")
ling_religious_terms <- read_csv("diversity_project - h2h3_categories.csv") %>%
  filter(category == "linguistic_religious")
national_terms <- read_csv("diversity_project - h2h3_categories.csv") %>%
  filter(category == "national")
directional_terms <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(classification == "directional")
south_american_ethnic_groups <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(classification == "south_america")
african_ethnic_groups <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(classification == "africa")
north_american_ethnic_groups <- read_csv("diversity_project - h2h3_categories.csv") %>%
  filter(classification == "north_america")
european_ethnic_groups <- read_csv("diversity_project - h2h3_categories.csv") %>%
  filter(classification == "europe")
asian_ethnic_groups <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(classification == "asia")
all_ethnic_groups <- read_csv("diversity_project - h2h3_categories.csv") %>% 
  filter(category == "ethnic")
# creating detection string patterns for all the categories
general_pop_terms <- paste(c("\\b(?i)(zxz", 
                     general_pop_terms$term, "zxz)\\b"), collapse = "|")
us_specific_terms <- paste(c("\\b(?i)(zxz", 
                     us_specific_terms$term, "zxz)\\b"), collapse = "|")
continental_terms <- paste(c("\\b(?i)(zxz", 
                     continental_terms$term, "zxz)\\b"), collapse = "|")
ling_religious_terms <- paste(c("\\b(?i)(zxz", 
                     ling_religious_terms$term, "zxz)\\b"), collapse = "|")
national_terms <- paste(c("\\b(?i)(zxz", 
                     national_terms$term, "zxz)\\b"), collapse = "|")
directional_terms <- paste(c("\\b(?i)(zxz", 
                     directional_terms$term, "zxz)\\b"), collapse = "|")
south_american_ethnic_groups <- paste(c("\\b(?i)(zxz", 
                     south_american_ethnic_groups$term, "zxz)\\b"), collapse = "|")
african_ethnic_groups <- paste(c("\\b(?i)(zxz", 
                     african_ethnic_groups$term, "zxz)\\b"), collapse = "|")
north_american_ethnic_groups <- paste(c("\\b(?i)(zxz", 
                     north_american_ethnic_groups$term, "zxz)\\b"), collapse = "|")
european_ethnic_groups <- paste(c("\\b(?i)(zxz", 
                     european_ethnic_groups$term, "zxz)\\b"), collapse = "|")
asian_ethnic_groups <- paste(c("\\b(?i)(zxz", 
                     asian_ethnic_groups$term, "zxz)\\b"), collapse = "|")
all_ethnic_groups <- paste(c("\\b(?i)(zxz", 
                     all_ethnic_groups$term, "zxz)\\b"), collapse = "|")

# graphing the population terms into various categories 
pop_terms_bycats <- abstract_data %>% 
  mutate(term_us = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = us_specific_terms), 
                       yes = "us-specific terms", no = word)) %>%
  mutate(term_cont = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = continental_terms), 
                       yes = "continental terms", no = word)) %>%
  mutate(term_nat = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = national_terms), 
                       yes = "national terms", no = word)) %>%
  mutate(term_direct = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = directional_terms), 
                       yes = "directional terms", no = word)) %>%
  mutate(term_genpop = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = general_pop_terms), 
                       yes = "general population labels", no = word)) %>%
  mutate(term_ethnic = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = all_ethnic_groups), 
                       yes = "ethnic groups", no = word)) %>% 
  mutate(term_pop = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = "\\b(?i)(population|populations)\\b"), 
                       yes = "population", no = word)) %>% 
  mutate(term_allpop = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = all_pop_terms), 
                       yes = "all population terms", no = word)) %>% 
  mutate(term_re = ifelse(test = str_detect(string = abstract_data$word, 
                      pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$race_ethnicity), "zqx)\\b"), collapse = "|")), 
                      yes = "race/ethnicity", no = word)) %>% 
  mutate(term_div = ifelse(test = str_detect(string = abstract_data$word, 
                      pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$diversity), "zqx)\\b"), collapse = "|")), 
                      yes = "diversity", no = word))
  
figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_us, sort = TRUE) %>% 
  filter(term_us == "us-specific terms") %>% 
  rename(term = term_us)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_cont, sort = TRUE) %>% 
  filter(term_cont == "continental terms") %>% 
  rename(term = term_cont) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_nat, sort = TRUE) %>% 
  filter(term_nat == "national terms") %>% 
  rename(term = term_nat) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_direct, sort = TRUE) %>% 
  filter(term_direct == "directional terms") %>% 
  rename(term = term_direct) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_genpop, sort = TRUE) %>% 
  filter(term_genpop == "general population labels") %>% 
  rename(term = term_genpop) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_ethnic, sort = TRUE) %>% 
  filter(term_ethnic == "ethnic groups") %>% 
  rename(term = term_ethnic) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_pop, sort = TRUE) %>% 
  filter(term_pop == "population") %>% 
  rename(term = term_pop) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_allpop, sort = TRUE) %>% 
  filter(term_allpop == "all population terms") %>% 
  rename(term = term_allpop) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_re, sort = TRUE) %>% 
  filter(term_re == "race/ethnicity") %>% 
  rename(term = term_re) %>% 
  bind_rows(figure_3A_counts)

figure_3A_counts <- pop_terms_bycats %>% 
  group_by(year) %>% 
  count(term_div, sort = TRUE) %>% 
  filter(term_div == "diversity") %>% 
  rename(term = term_div) %>% 
  bind_rows(figure_3A_counts)
```

```{r figure 3a, echo=FALSE}
figure_3A <- figure_3A_counts %>%   
  filter(term != "all population terms") %>% 
  ggplot() + geom_line(aes(y = n, x = year, colour = term), size = 1.3, stat="identity") + 
  theme_minimal() +
  labs(title = "     Figure 3A. Total Growth of \n Separated Diversity Terms (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("continental terms", "directional terms", "diversity", "ethnic terms",
  "general population labels", "national terms", "population", "race/ethnicity", "us-specific terms"),
                     values=c("#3B9AB2", "#1b7524", "#E21F00", "#d23ccc", 
                              "#61B329", "#E86F00", "#E2B306", "#00468b", "#6b0f6b" )) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)); ggplotly() # pop green, race yellow 
```

In support of Fujimura and Rajagopalan (2011), Figures 3A & 3B reveal that continental and country terms outgrew the use of racial, ethnic and US-Specific terminology over the past three decades. More specifically, Figure 3A shows that country terms have exhibited a 35-fold increase followed closely by the use of continental terms. Directional terms, which are both included in the broader continental classification (as they are typically used to refer to specific areas within continents) and broken apart into their own independent line, have also grown notably, but pale in comparison with the rise in continent and country terms. There has also been a noticable uptick in general population terms, which is a bit of catch-all category. We can see that the term population, which is included both in the general population classification in addition to being plotted independently, accounts for more than half of this growth while race and ethnicity makes up considerably less of that category's ascent. Ethnic terms, in contrast, have only increased marginally, growing at roughly the same rate as US-specific terminology since 1990. 

```{r figure 3b, echo=FALSE}
pop_terms_bycats <- pop_terms_bycats %>% 
  mutate(continental_terms = ifelse(test = str_detect(string = term_cont, 
                       pattern = "continental terms"), yes = 1, no = 0)) %>%
  mutate(directional_terms = ifelse(test = str_detect(string = term_direct, 
                       pattern = "directional terms"), yes = 1, no = 0)) %>%
  mutate(diversity = ifelse(test = str_detect(string = term_div, 
                       pattern = "diversity"), yes = 1, no = 0)) %>%
  mutate(ethnic_terms = ifelse(test = str_detect(string = term_ethnic, 
                       pattern = "ethnic groups"), yes = 1, no = 0)) %>%
  mutate(gen_pop_terms = ifelse(test = str_detect(string = term_genpop, 
                       pattern = "general population labels"), yes = 1, no = 0)) %>%
  mutate(national_terms = ifelse(test = str_detect(string = term_nat, 
                       pattern = "national terms"), yes = 1, no = 0)) %>%
  mutate(population = ifelse(test = str_detect(string = term_pop, 
                       pattern = "\\b((?<!general )population)\\b"), yes = 1, no = 0)) %>%
  mutate(all_pop_terms = ifelse(test = str_detect(string = term_allpop, 
                       pattern = "all population terms"), yes = 1, no = 0)) %>% 
  mutate(race_ethnicity = ifelse(test = str_detect(string = term_re, 
                       pattern = "race/ethnicity"), yes = 1, no = 0)) %>% 
  mutate(us_terms = ifelse(test = str_detect(string = term_us, 
                       pattern = "us-specific terms"), yes = 1, no = 0)) 

# total articles each year 
bycats_prc_counts <- text_data %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n) 
# articles with term mentioned each year 
bycats_prc_counts <- pop_terms_bycats %>% 
  filter(gen_pop_terms == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_general_terms = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_general_terms = round(cnt_general_terms / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% 
  filter(us_terms == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_us_terms = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_us_terms = round(cnt_us_terms / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(continental_terms == 1) %>% 
  group_by(year) %>% summarise(cnt_continental_terms = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_continental_terms = round(cnt_continental_terms / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(national_terms == 1) %>% 
  group_by(year) %>% summarise(cnt_national_terms = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_national_terms = round(cnt_national_terms / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(ethnic_terms == 1) %>% 
  group_by(year) %>% summarise(cnt_ethnic_terms = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_ethnic_terms = round(cnt_ethnic_terms / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(population == 1) %>% 
  group_by(year) %>% summarise(cnt_population = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_population = round(cnt_population / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(directional_terms == 1) %>% 
  group_by(year) %>% summarise(cnt_directional_terms = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_directional_terms = round(cnt_directional_terms / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(diversity == 1) %>% 
  group_by(year) %>% summarise(cnt_diversity = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2))
bycats_prc_counts <- pop_terms_bycats %>% filter(race_ethnicity == 1) %>% 
  group_by(year) %>% summarise(cnt_race_ethnicity = n_distinct(pubmed_id)) %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(prc_race_ethnicity = round(cnt_race_ethnicity / total * 100, digits = 2))

figure_3B <- bycats_prc_counts %>% ggplot() + 
  geom_line(aes(y = prc_continental_terms, x = year, colour = "continental terms"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_directional_terms, x = year, colour = "directional terms"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_ethnic_terms, x = year, colour = "ethnic terms"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_general_terms, x = year, colour = "general population labels"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_national_terms, x = year, colour = "national terms"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_population, x = year, colour = "population"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_race_ethnicity, x = year, colour = "race/ethnicity"), size=1.3, stat="identity") +
  geom_line(aes(y = prc_us_terms, x = year, colour = "us-specific terms"), size=1.3, stat="identity") +
  labs(title = "    Figure 3B. Proportional Growth of \n Separated Diversity Terms (1990-2017)",
       color =  "Term") + 
  theme_minimal() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("continental terms", "directional terms", "diversity", "ethnic terms",
  "general population labels", "national terms", "population", "race/ethnicity", "us-specific terms"),
                     values=c("#3B9AB2", "#1b7524", "#E21F00", "#d23ccc", 
                              "#61B329", "#E86F00", "#E2B306", "#00468b", "#6b0f6b" )) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)); ggplotly()
```

Figure 3B displays the proportional changes in the use of population terminology. Again, country terms have grown the most; surging from 12% of abstracts to now being mentioned in 35% of abstracts. Continental terms have also risen by about 15% over that time period. General population terms, the term population, and directional terms follow analogous patterns, rising 10%, 8% and 6% respectively since 1990. In contrast, there has been almost no variation in the use of racial, ethnic or US-specific terminology over the past three decades. While ethnic terms has stayed consistently low across our analysis period, race/ethnicity and US-specific terminology followed a parabolic curve that increases up until the mid-2000s before declining back to its baseline in 2017.  

### Main Takeaway 

Together, these analyses reveal that national and continental terms outgrew the use of racial, ethnic and US-Specific - a trend that holds for both the raw counts and proportions in Figures 3A & 3B. These findings strongly support Fujimura and Rajagopalan’s (2011) thesis that national and continental terms are now more commonly used than racial and ethnic terms in the scientific literature, suggesting that scholars of diversity must be attentive to the emerging strategies that biomedical researchers use to enact population differences in their research. Focusing solely on discourses related to race and ethnicity likely fails to capture the more subtle ways that biomedical scholars implemented new and potentially dangerous forms of population difference in their research.  

### Appendix 

Here is a list of the 2,200+ terms that were collapsed into the "all population terms" category in these analyses. You can use the search box to find a specific term you are interested in. Please note that this dictionary is still in progress and the sub/categories are both fluid and imperfect. If you have suggestions about what to add and/or reclassify, please send an email to the authors using the email link in the top-right corner.

```{r, warning=FALSE, message=FALSE, echo = FALSE}
setwd("~/Documents/Diversity/Data")
all_pop_terms_table <- read_csv("diversity_project - h2h3_categories.csv")
all_pop_terms_table <- all_pop_terms_table #%>% select(term, category)
all_pop_terms_table[is.na(all_pop_terms_table)] <- ""
DT::datatable(all_pop_terms_table)
```
