==> src/02_text_trends/00_antiquated/09_final-figures-new.Rmd <==
---
title: "Publication Figures"
author: "Brandon L. Kramer"
date: "1/18/2020"
output: html_document
---

```{r setup, include = FALSE}
rm(list = ls())
knitr::opts_knit$set(root.dir = "~/Documents/Diversity/Data")
#knitr::opts_knit$set(root.dir = "C:/Users/bkram/CloudStation/Diversity/Data")
#knitr::opts_knit$set(root.dir = "~/data/diversity-data")
# loading packages 
for (pkg in c("tidyverse", "igraph", "tidytext", "ggplot2", 
              "grid", "gridExtra", "reshape2", "extrafont",
               "plotly", "stringi", "networkD3")) {library(pkg, character.only = TRUE)}
```

```{r supp fig 1, fig.width=7, fig.height=4}

diversity_corpus <- read_csv("diversity_corpus.csv") %>% 
  #drop_na(abstract) %>%
  filter(year != "2018")

biomedical_corpus <- read_csv("biomedical_corpus.csv") %>% 
  rowid_to_column(var = "id") %>% 
  rename(authors = AU, title = TI, publication = SO, #author_keywords = DE,
         abstract = AB, references = CR, year = PY, times_cited = TC) %>% 
  select(id, authors, title, year, publication, abstract, references, year, times_cited) %>% 
  #drop_na(abstract) %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018")

dv_by_year <- diversity_corpus %>% 
  group_by(year) %>% 
  count() %>% 
  rename(diversity = n)

bm_by_year <- biomedical_corpus %>% 
  group_by(year) %>% 
  count() %>% 
  rename(biomedical = n)

dv_by_year <- melt(dv_by_year, "year")
bm_by_year <- melt(bm_by_year, "year")

by_year <- rbind(dv_by_year, bm_by_year)
by_year <- by_year %>% 
  rename(domain = variable, total = value) %>% 
  mutate(year = as.factor(year),
         domain = as.factor(domain),
         total = as.numeric(total)) %>%
  arrange(year)

ggplot(data = by_year, aes(x = year, y = total, fill = domain)) + 
  geom_bar(stat="identity", position="dodge") +
  labs(title = "Supplementary Figure 1. Number of Publication Abstracts in Datasets (1990-2017)",
       caption = "Data Source: ISI Web of Science") + 
  theme(axis.title.x = element_blank(),
        plot.title = element_text(size=11, hjust = 0.2, face="bold"),
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        panel.background = element_rect(fill = "white")) +
  labs(fill = "Sample") +
  ylab("Number of Abstracts") +
  scale_fill_manual(values=c('black','darkgray')) +
  scale_x_discrete(labels = c("1991" = "", "1992" = "", "1993" = "", "1994" = "",
                              "1996" = "", "1997" = "", "1998" = "", "1999" = "",
                              "2001" = "", "2002" = "", "2003" = "", "2004" = "",
                              "2006" = "", "2007" = "", "2008" = "", "2009" = "",
                              "2011" = "", "2012" = "", "2013" = "", "2014" = "",
                              "2016" = "", "2017" = ""))
  
```

```{r figure 1ab, fig.height=5, fig.width=11.5, echo=FALSE}

for (pkg in c("tidyverse", "igraph", "tidytext", "ggplot2", "DT",
              "grid", "gridExtra", "reshape2", "extrafont", "data.table", "maditr",
              "plotly", "stringi", "networkD3")) {library(pkg, character.only = TRUE)}

setwd("~/Documents/Diversity/Data")
# loading the .csv file 
biomed_text_data <- read_csv("biomedical_corpus.csv") %>% 
  rowid_to_column(var = "id") %>% 
  rename(authors = AU, title = TI, publication = SO, #author_keywords = DE,
         abstract = AB, references = CR, year = PY, times_cited = TC) %>% 
  select(id, authors, title, year, publication, abstract, references, year, times_cited) %>% 
  drop_na(abstract) %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018") 

# tokenizing the abstract data into words 
biomed_abstract_data <- biomed_text_data %>% 
  unnest_tokens(word, abstract) #%>% anti_join(stop_words)

# adding custom set of stopwords 
my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier", 
                                "women.recommendations", 
                                "cs.man.ac.uk", "sexes.please", "movements.baby"))

biomed_abstract_data <- biomed_abstract_data %>% 
  anti_join(my_stopwords)

setwd("~/Documents/Diversity/Data")
divictionary <- read_csv("diversity_project - h1_dictionary.csv") 

general_pop_terms <- biomed_abstract_data %>% 
  as.data.table() %>%
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$aging), "zqx)\\b"), collapse = "|")), 
         yes = "aging", no = word)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$ancestry), "zqx)\\b"), collapse = "|")), 
         yes = "ancestry", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$cultural), "zqx)\\b"), collapse = "|")), 
         yes = "cultural", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$diversity), "zqx)\\b"), collapse = "|")), 
         yes = "diversity", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$genetic), "zqx)\\b"), collapse = "|")), 
         yes = "genetic", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$social_class), "zqx)\\b"), collapse = "|")), 
         yes = "socio-economic", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$minority), "zqx)\\b"), collapse = "|")), 
         yes = "minority", no = term)) %>%
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$population), "zqx)\\b"), collapse = "|")), 
         yes = "population", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$race_ethnicity), "zqx)\\b"), collapse = "|")), 
         yes = "race/ethnicity", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$sex_gender), "zqx)\\b"), collapse = "|")), 
         yes = "sex/gender", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$sexuality), "zqx)\\b"), collapse = "|")), 
         yes = "sexuality", no = term)) %>% 
  dt_mutate(di_aging = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(aging)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_ancestry = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(ancestry)\\b"), yes = 1, no = 0)) %>% 
  dt_mutate(di_cultural = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(cultural)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_class = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(socio-economic)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_diversity = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(diversity)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_genetic = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(genetic)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_minority = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(minority)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_population = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(population)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_racial = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(race/ethnicity)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_sexgender = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(sex/gender)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_sexuality = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(sexuality)\\b"), yes = 1, no = 0)) 

# total articles each year 
gen_pop_prc_counts <- biomed_text_data %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n) 

# articles with term mentioned each year 
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_racial == 1) %>% #racial 
  group_by(year) %>% 
  summarise(cnt_racial = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>%  
  mutate(prc_racial = round(cnt_racial / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_cultural == 1) %>% #cultural
  group_by(year) %>% 
  summarise(cnt_cultural = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_cultural = round(cnt_cultural / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_population == 1) %>% #population
  group_by(year) %>% 
  summarise(cnt_population = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_population = round(cnt_population / total * 100, digits = 2)) 
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_ancestry == 1) %>% # ancestry
  group_by(year) %>% 
  summarise(cnt_ancestry = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_ancestry = round(cnt_ancestry / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_diversity == 1) %>% #diversity
  group_by(year) %>% 
  summarise(cnt_diversity = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_genetic == 1) %>% #genetic
  group_by(year) %>% 
  summarise(cnt_genetic = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_genetic = round(cnt_genetic / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_sexgender == 1) %>% #sex/gender
  group_by(year) %>% 
  summarise(cnt_sexgender = n_distinct(id)) %>% 
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_sexgender = round(cnt_sexgender / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_sexuality == 1) %>% # sexuality
  group_by(year) %>% 
  summarise(cnt_sexuality = n_distinct(id)) %>%  
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_sexuality = round(cnt_sexuality / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_aging == 1) %>% # aging
  group_by(year) %>% 
  summarise(cnt_aging = n_distinct(id)) %>%  
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_aging = round(cnt_aging / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_class == 1) %>% # class
  group_by(year) %>% 
  summarise(cnt_class = n_distinct(id)) %>%  
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_class = round(cnt_class / total * 100, digits = 2))
gen_pop_prc_counts <- general_pop_terms %>% 
  filter(di_minority == 1) %>% # minority
  group_by(year) %>% 
  summarise(cnt_minority = n_distinct(id)) %>%  
  right_join(gen_pop_prc_counts, by = "year") %>% 
  mutate(prc_minority = round(cnt_minority / total * 100, digits = 2))

#graphs 
figure_1A <- general_pop_terms %>% 
  # grouping by year and filter relevant population terms 
  group_by(year) %>% count(term, sort = TRUE) %>% ungroup() %>% 
  filter(term == "diversity" | term == "cultural" | #term == "genetic" | term == "ancestry" | 
           term == "sex/gender" | term == "sexuality" | term == "aging" | term == "minority" |
           term == "race/ethnicity" | term == "population" | term == "socio-economic") %>% 
  # graphing terms over time 
  ggplot() + 
  geom_line(aes(y = n, x = year, colour = term), size = 1.3, stat="identity") + 
  theme_bw() +
  labs(title = "     Figure 1A. Total Growth of Diversity-Related \n Terminology in Diversity Sample (1990-2017)",
       color =  "Term") + 
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  #scale_color_manual(labels=c("aging",  "ancestry", "cultural", "diversity","genetic", "minority",
  #                            "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
  #                   values=c("#f2be00", "#000000",  "#61B329", "#3B9AB2", "#E86F00","#d23ccc", 
  #                            "#1b7524", "#005BB7", "#CC0000", "#ff8396", "#800080")) +
  scale_color_manual(labels=c("aging", "cultural", "diversity", "minority",
                              "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
                     values=c("#E86F00",  "#61B329", "#000000", "#d23ccc", 
                              "#f2be00", "#3B9AB2", "#CC0000", "#ff8396", "#800080")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)) 
  #scale_color_manual( values=c( "black", "#666666", "grey","black", "#666666", "grey"))

# graphing terms over time 
figure_1B <- gen_pop_prc_counts %>% ggplot() + 
  geom_line(aes(y = prc_aging, x = year, colour = "aging"), size = 1.3, stat="identity") +
  #geom_line(aes(y = prc_ancestry, x = year, colour = "ancestry"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_class, x = year, colour = "socio-economic"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_cultural, x = year, colour = "cultural"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_ethnic, x = year, colour = "ethnic"), size = 1.3, stat="identity") +
  #geom_line(aes(y = prc_genetic, x = year, colour = "genetic"), size = 1.3, stat="identity") +
  #geom_line(aes(y = prc_linguistic, x = year, colour = "linguistic"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_minority, x = year, colour = "minority"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_population, x = year, colour = "population"), size = 1.3, stat="identity") +
  #geom_line(aes(y = prc_racial, x = year, colour = "racial"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race/ethnicity"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_sexgender, x = year, colour = "sex/gender"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_sexuality, x = year, colour = "sexuality"), size = 1.3, stat="identity") +
  labs(title = "    Figure 1B. Proportional Growth in Diversity-Related \n Terminology in Diversity Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        #legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  #scale_color_manual(labels=c("aging",  "ancestry", "cultural", "diversity","genetic", "minority",
  #                            "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
  #                   values=c("#f2be00", "#000000",  "#61B329", "#3B9AB2", "#E86F00", "#d23ccc", 
  #                            "#1b7524", "#005BB7", "#CC0000", "#ff8396", "#800080")) +
  scale_color_manual(labels=c("aging", "cultural", "diversity", "minority",
                              "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
                     values=c("#E86F00",  "#61B329", "#000000", "#d23ccc", 
                              "#f2be00", "#3B9AB2", "#CC0000", "#ff8396", "#800080")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)) 
  #scale_color_manual( values=c( "black", "#666666", "grey","black", "#666666", "grey"))

custom_variables <- c("prc_ancestry", "prc_cultural", "prc_diversity", 
                      "prc_genetic", "prc_population", "prc_racial", "prc_sexgender", "prc_sexuality")
custom_linetype <- c("solid", "solid",  "solid", "twodash", "dotted", "dashed")

legend <- gen_pop_prc_counts %>% ggplot() + 
  geom_line(aes(y = prc_aging, x = year, colour = "aging"), 
            size = 1.5, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_ancestry, x = year, colour = "ancestry"), 
  #          size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_cultural, x = year, colour = "cultural"), 
            size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), 
            size = 1.5, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_genetic, x = year, colour = "genetic"), 
  #          size = 1.5, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_linguistic, x = year, colour = "linguistic"), 
  #          size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_minority, x = year, colour = "minority"), 
            size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_population, x = year, colour = "population"), 
            size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_racial, x = year, colour = "race/ethnicity"), 
            size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_sexgender, x = year, colour = "sex/gender"), 
            size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_sexuality, x = year, colour = "sexuality"), 
            size = 1.5, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_class, x = year, colour = "socio-economic"), size = 1.3, stat="identity") +
  labs(title = "    Figure 1B. Proportional Growth in Diversity-Related \n Terminology in Diversity Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "bottom",
        #legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  #scale_color_manual(labels=c("aging",  "ancestry", "cultural", "diversity","genetic", "minority",
  #                            "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
  #                   values=c("#f2be00", "#000000",  "#61B329", "#3B9AB2", "#E86F00", "#d23ccc", 
  #                            "#1b7524", "#005BB7", "#CC0000", "#ff8396", "#800080")) +
  scale_color_manual(labels=c("aging", "cultural", "diversity", "minority",
                              "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
                     values=c("#E86F00",  "#61B329", "#000000", "#d23ccc", 
                              "#f2be00", "#3B9AB2", "#CC0000", "#ff8396", "#800080")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))
  #scale_color_manual( values=c( "black", "#666666", "grey","black", "#666666", "grey")) 
  #scale_linetype_manual(name = custom_variables, values = custom_linetype)
legend <- cowplot::get_legend(legend)

grid.arrange(figure_1A, figure_1B, 
             legend, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )

```

```{r epa figure 1ab, eval=FALSE, echo=FALSE}
ggplotly(figure_1A)
ggplotly(figure_1B)

# raw count growth 
# diversity
4574/380
# population 
1832/145
# genetic
1191/77
# sex/gender
729/33
# aging
593/28
# cultural
478/38
# race/ethnicity 
342/16
```

```{r figure 1cd, fig.height=5, fig.width=11.5, echo=FALSE}

setwd("~/Documents/Diversity/Data")
# loading the .csv file 
text_data <- read_csv("diversity_corpus.csv") %>% 
  drop_na(abstract) %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018") 
# tokenizing the abstract data into words 
abstract_data <- text_data %>% 
  unnest_tokens(word, abstract) %>% 
  anti_join(stop_words)

my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier", 
                                "women.recommendations", 
                                "cs.man.ac.uk", "sexes.please", "movements.baby"))

abstract_data <- abstract_data %>% 
  anti_join(my_stopwords)

# lets draw all of our strings from the diversity_dictionary (divictionary)
setwd("~/Documents/Diversity/Data")
divictionary <- read_csv("diversity_project - h1_dictionary.csv") 

# recoding race*, ethnicity* and cultural* terminology
general_pop_terms_div <- abstract_data %>% 
  as.data.table() %>%
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$aging), "zqx)\\b"), collapse = "|")), 
         yes = "aging", no = word)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$ancestry), "zqx)\\b"), collapse = "|")), 
         yes = "ancestry", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$cultural), "zqx)\\b"), collapse = "|")), 
         yes = "cultural", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$diversity), "zqx)\\b"), collapse = "|")), 
         yes = "diversity", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$genetic), "zqx)\\b"), collapse = "|")), 
         yes = "genetic", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$minority), "zqx)\\b"), collapse = "|")), 
         yes = "minority", no = term)) %>%
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$population), "zqx)\\b"), collapse = "|")), 
         yes = "population", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$race_ethnicity), "zqx)\\b"), collapse = "|")), 
         yes = "race/ethnicity", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$sex_gender), "zqx)\\b"), collapse = "|")), 
         yes = "sex/gender", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$sexuality), "zqx)\\b"), collapse = "|")), 
         yes = "sexuality", no = term)) %>% 
  dt_mutate(term = ifelse(test = str_detect(string = word, 
         pattern = paste(c("\\b(?i)(zqx", na.omit(divictionary$social_class), "zqx)\\b"), collapse = "|")), 
         yes = "socio-economic", no = term)) %>% 
  dt_mutate(di_aging = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(aging)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_ancestry = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(ancestry)\\b"), yes = 1, no = 0)) %>% 
  dt_mutate(di_cultural = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(cultural)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_diversity = ifelse(test = str_detect(string = term, 
                       pattern = "diversity"), yes = 1, no = 0)) %>%
  dt_mutate(di_genetic = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(genetic)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_minority = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(minority)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_population = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(population)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_racial = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(race/ethnicity)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_sexgender = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(sex/gender)\\b"), yes = 1, no = 0)) %>%
  dt_mutate(di_sexuality = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(sexuality)\\b"), yes = 1, no = 0)) %>% 
  dt_mutate(di_class = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(socio-economic)\\b"), yes = 1, no = 0)) 

# total articles each year 
gen_pop_prc_counts_div <- text_data %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n) 
# articles with term mentioned each year 
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_racial == 1) %>% 
  group_by(year) %>% summarise(cnt_racial = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>%  
  mutate(prc_racial = round(cnt_racial / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_cultural == 1) %>% 
  group_by(year) %>% summarise(cnt_cultural = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_cultural = round(cnt_cultural / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_population == 1) %>% 
  group_by(year) %>% summarise(cnt_population = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_population = round(cnt_population / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_ancestry == 1) %>% 
  group_by(year) %>% summarise(cnt_ancestry = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_ancestry = round(cnt_ancestry / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_diversity == 1) %>% 
  group_by(year) %>% summarise(cnt_diversity = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_genetic == 1) %>% 
  group_by(year) %>% summarise(cnt_genetic = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_genetic = round(cnt_genetic / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_sexgender == 1) %>% 
  group_by(year) %>% summarise(cnt_sexgender = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_sexgender = round(cnt_sexgender / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_sexuality == 1) %>% 
  group_by(year) %>% summarise(cnt_sexuality = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_sexuality = round(cnt_sexuality / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_aging == 1) %>% 
  group_by(year) %>% summarise(cnt_aging = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_aging = round(cnt_aging / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_minority == 1) %>% 
  group_by(year) %>% summarise(cnt_minority = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_minority = round(cnt_minority / total * 100, digits = 2))
gen_pop_prc_counts_div <- general_pop_terms_div %>% filter(di_class == 1) %>% 
  group_by(year) %>% summarise(cnt_class = n_distinct(pubmed_id)) %>% 
  right_join(gen_pop_prc_counts_div, by = "year") %>% 
  mutate(prc_class = round(cnt_class / total * 100, digits = 2))

figure_1C <- gen_pop_prc_counts_div %>% 
  ggplot() + 
  geom_line(aes(y = cnt_aging, x = year, colour = "aging"), 
            size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = cnt_ancestry, x = year, colour = "ancestry"), 
  #          size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = cnt_cultural, x = year, colour = "cultural"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = cnt_diversity, x = year, colour = "diversity"), 
            size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = cnt_genetic, x = year, colour = "genetic"), 
  #          size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = cnt_lingustic, x = year, colour = "linguistic"), 
  #          size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = cnt_minority, x = year, colour = "minority"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = cnt_population, x = year, colour = "population"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = cnt_racial, x = year, colour = "race/ethnicity"), 
            size = 1.3, stat="identity", linetype = "solid") + 
  geom_line(aes(y = cnt_sexgender, x = year, colour = "sex/gender"), 
            size = 1.3, stat="identity", linetype = "solid") + 
  geom_line(aes(y = cnt_sexuality, x = year, colour = "sexuality"), 
            size = 1.3, stat="identity", linetype = "solid") + 
  geom_line(aes(y = cnt_class, x = year, colour = "socio-economic"), 
            size = 1.3, stat="identity", linetype = "solid") +
  theme_bw() +
  labs(title = "     Figure 1C. Total Growth of Diversity-Related \n Terminology in Biomedical Sample (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  #scale_color_manual(labels=c("aging",  "ancestry", "cultural", "diversity","genetic", "minority",
  #                            "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
  #                   values=c("#f2be00", "#000000",  "#61B329", "#3B9AB2", "#E86F00", "#d23ccc", 
  #                            "#1b7524", "#005BB7", "#CC0000", "#ff8396", "#800080")) +
  scale_color_manual(labels=c("aging", "cultural", "diversity", "minority",
                              "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
                     values=c("#E86F00",  "#61B329", "#000000", "#d23ccc", 
                              "#f2be00", "#3B9AB2", "#CC0000", "#ff8396", "#800080")) +
  scale_y_log10(breaks = c(0, 5, 10, 25, 50, 100, 250, 500, 1000, 2500, 5000)) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)) 
  #scale_linetype_identity( linetype=c("solid", "solid",  "solid", "twodash", "dotted", "dashed")) + 
  #scale_color_manual( values=c( "black", "#666666", "grey","black", "#666666", "grey"))

# graphing terms over time 
figure_1D <- gen_pop_prc_counts_div %>% ggplot() + 
  geom_line(aes(y = prc_aging, x = year, colour = "aging"), 
            size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_ancestry, x = year, colour = "ancestry"), 
  #          size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_cultural, x = year, colour = "cultural"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), 
            size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_genetic, x = year, colour = "genetic"), 
  #          size = 1.3, stat="identity", linetype = "solid") +
  #geom_line(aes(y = prc_linguistic, x = year, colour = "linguistic"), 
  #          size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_minority, x = year, colour = "minority"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_population, x = year, colour = "population"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_racial, x = year, colour = "race/ethnicity"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_sexgender, x = year, colour = "sex/gender"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_sexuality, x = year, colour = "sexuality"), 
            size = 1.3, stat="identity", linetype = "solid") +
  geom_line(aes(y = prc_class, x = year, colour = "socio-economic"), size = 1.3, stat="identity") +
  labs(title = "    Figure 1D. Proportional Growth Diversity-Related \n Terminology in Biomedical Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        #legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  #scale_color_manual(labels=c("aging",  "ancestry", "cultural", "diversity","genetic", "minority",
  #                            "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
  #                   values=c("#f2be00", "#000000",  "#61B329", "#3B9AB2", "#E86F00", "#d23ccc", 
  #                            "#1b7524", "#005BB7", "#CC0000", "#ff8396", "#800080")) +
  scale_color_manual(labels=c("aging", "cultural", "diversity", "minority",
                              "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
                     values=c("#E86F00",  "#61B329", "#000000", "#d23ccc", 
                              "#f2be00", "#3B9AB2", "#CC0000", "#ff8396", "#800080")) +
  scale_y_log10(breaks = c(1, 2.5, 5, 10, 25, 50, 100)) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)) 
  #scale_linetype_manual( linetype=c("solid", "solid",  "solid", "twodash", "dotted", "dashed")) +
  #scale_color_manual( values=c( "black", "#666666", "grey","black", "#666666", "grey")) 

legend <- gen_pop_prc_counts %>% ggplot() + 
  geom_line(aes(y = prc_aging, x = year, colour = "aging"), size = 1.5, stat="identity") +
  #geom_line(aes(y = prc_ancestry, x = year, colour = "ancestry"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_cultural, x = year, colour = "cultural"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size = 1.5, stat="identity", 
            linetype = "solid") +
  #geom_line(aes(y = prc_ethnic, x = year, colour = "ethnic"), size = 1, stat="identity") +
  #geom_line(aes(y = prc_genetic, x = year, colour = "genetic"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_minority, x = year, colour = "minority"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_population, x = year, colour = "population"), size = 1.5, stat="identity") +
  #geom_line(aes(y = prc_racial, x = year, colour = "racial"), size = 1, stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race/ethnicity"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_sexgender, x = year, colour = "sex/gender"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_sexuality, x = year, colour = "sexuality"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_class, x = year, colour = "socio-economic"), size = 1.3, stat="identity") +
  labs(title = "    Figure 1D. Proportional Growth Diversity-Related \n Terminology in Biomedical Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  #scale_color_manual(labels=c("aging",  "ancestry", "cultural", "diversity","genetic", "minority",
  #                            "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
  #                   values=c("#f2be00", "#000000",  "#61B329", "#3B9AB2", "#E86F00", "#d23ccc", 
  #                            "#1b7524", "#005BB7", "#CC0000", "#ff8396", "#800080")) +
  scale_color_manual(labels=c("aging", "cultural", "diversity", "minority",
                              "population","race/ethnicity","sex/gender","sexuality","socio-economic"),
                     values=c("#E86F00",  "#61B329", "#000000", "#d23ccc", 
                              "#f2be00", "#3B9AB2", "#CC0000", "#ff8396", "#800080")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)) 
legend <- cowplot::get_legend(legend)

grid.arrange(figure_1C, figure_1D, 
             legend, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             heights = c(2.5, 0.4)
             )
```

```{r epa figures 1cd, fig.height=7, fig.width=11.5, echo=FALSE, eval=FALSE}
#sex/gender
1558/35
#genetic
1578/46
#population 
814/22
#aging 
880/50
#cultural 
521/18
#diversity
381/7
#race/ethnicity 
212/5
#minority
160/3
ggplotly(figure_1C)
ggplotly(figure_1D)
```

```{r binding 1ab + 1cd, fig.height=7, fig.width=11.5, echo=FALSE}
lay <- rbind(c(1,2),
             c(3,4),
             c(5,5))


grid.arrange(figure_1A, figure_1B, 
             figure_1C, figure_1D,
             legend, 
             nrow = 3, ncol=2,
             layout_matrix = lay,
             #widths = c(2.7, 2.7), 
             heights = c(4.2, 4.2, 0.8)
             )
```

```{r figures_2AB, fig.height=5, fig.width=11.5, message = FALSE, results = FALSE, warning = FALSE, echo=FALSE}
setwd("~/Documents/Diversity/Data")
bm_text_data <- read_csv("biomedical_corpus.csv") %>% 
  rowid_to_column(var = "id") %>% 
  rename(authors = AU, title = TI, publication = SO, #author_keywords = DE,
         abstract = AB, references = CR, year = PY, times_cited = TC, 
         department = C1, pubmed_id = UT) %>% 
  #drop_na(abstract) %>%
  select(id, authors, title, year, publication, abstract, 
         references, year, times_cited, pubmed_id, department) %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018")

# tokenizing the abstract data into words 
bm_abstract_data <- bm_text_data %>% 
  unnest_tokens(word, abstract) #%>% anti_join(stop_words)
# adding custom set of stopwords 
my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier", 
                                "women.recommendations", 
                                "cs.man.ac.uk", "sexes.please", "movements.baby"))
bm_abstract_data <- bm_abstract_data %>% 
  anti_join(stop_words) %>% 
  anti_join(my_stopwords); rm(my_stopwords)

black_bigrams <- "\\b(?i)(african american|african americans|afro american|afro americans|african-american|african-americans|afro-american|afro-americans)\\b"
native_american_bigrams <- "\\b(?i)(native american|native americans|american indian|american indians|alaskan native|alaskan natives|alaska nativez|alaska natives|native-american|native-americans|american-indian|american-indians|alaskan-native|alaskan-natives|alaska-native|alaska-natives)\\b"
pacific_islander_bigrams <- "\\b(?i)(pacific islander|pacific islanders|native hawaiian|native hawaiians|pacific-islander|pacific-islanders|native-hawaiian|native-hawaiians)\\b"
hispanic_american_bigrams <- "\\b(?i)(hispanic-american|hispanic american|hispanic-americans|hispanic americans)\\b"
asian_american_bigrams <- "\\b(?i)(asian-american|asian american)\\b"
asian_american_bigrams <- "\\b(?i)(caucasian-american|caucasian american)\\b"

setwd("~/Documents/Diversity/Data")
h1_dictionary <- read_csv("diversity_project - h1_dictionary.csv")
h3_dictionary <- read_csv("diversity_project - h3_dictionary.csv")
omb_black <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$black), "z3x)\\b"), collapse = "|")
omb_native_american <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$native_american), "z3x)\\b"), collapse = "|")
omb_pacific_islander <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$pacific_islander), "z3x)\\b"), collapse = "|")
omb_hispanic_latinx <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$hispanic_latinx), "z3x)\\b"), collapse = "|")
omb_asian <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$asian), "z3x)\\b"), collapse = "|")
omb_white <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$white), "z3x)\\b"), collapse = "|")
omb_racial <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$race), "z3x)\\b"), collapse = "|")
omb_ethnicity <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$ethnicity), 
                         "z3x)\\b"), collapse = "|")
omb_diversity <- paste(c("\\b(?i)(z3x",na.omit(h1_dictionary$diversity), 
                         "z3x)\\b"), collapse = "|")

# recoding frequently occuring population-related bigrams 
us_pop_bigrams <- bm_text_data %>% 
  unnest_tokens(bigram, abstract, token = "ngrams", n = 2) %>% 
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = black_bigrams), yes = "african-american", no = bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = native_american_bigrams), yes = "native-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = pacific_islander_bigrams), yes = "pacific-islander", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = hispanic_american_bigrams), yes = "hispanic-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = asian_american_bigrams), yes = "asian-american", no = rc_bigram)) %>%
  select(-bigram) %>% 
  filter(rc_bigram == "african-american" | 
         rc_bigram == "native-american" | 
         rc_bigram == "hispanic-american" |
         rc_bigram == "asian-american" |
         rc_bigram == "pacific-islander") %>% 
  count(rc_bigram, year, sort = TRUE) %>% 
  rename(word = rc_bigram)

# combining the hyphenated words to our full dataset 
us_pop_counts <- bm_abstract_data %>% 
  group_by(year) %>% 
  count(word, sort = TRUE) %>% 
  ungroup() %>% 
  select(word, year, n) %>% 
  bind_rows(us_pop_bigrams) %>% 
  arrange(-n)

us_pop_counts <- us_pop_counts %>% 
  mutate(term = ifelse(test = str_detect(string = word,
                       pattern = omb_black), yes = "black", no = word)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_white), yes = "white", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_asian), yes = "asian", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                      pattern = omb_hispanic_latinx), yes = "hispanic/latinx", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_racial), yes = "race", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_ethnicity), yes = "ethnicity", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_native_american), yes = "native-american", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_pacific_islander), yes = "pacific-islander", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_diversity), yes = "diversity", no = term)) %>% 
  filter(term == "black" | term == "white" | term == "asian" |
         term == "hispanic/latinx" | term == "race" | term == "ethnicity" |
         term == "native-american" | term == "pacific-islander" | term == "diversity" ) %>%  
  group_by(term, year) %>% 
  summarise(n = sum(n)) %>% 
  arrange(-n)

# graphing us-specific terms over time using census categories 
# https://www.census.gov/topics/population/race/about.html
figure_2A <- ggplot(us_pop_counts) + 
  geom_line(aes(y = n, x = year, colour = term), 
            size = 1.5, stat="identity") + 
  theme_bw() +
  labs(title = "     Figure 2A. Total Growth of \n US-Specific Terms in Biomedical Sample (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(labels=c("asian","black","diversity","ethnicity","hispanic/latinx",
                              "native-american","pacific-islander","race","white"),
                     values=c("#E86F00", "#3B9AB2", "#000000", "#E21F00", "#00468b", 
                              "#6b0f6b", "#d23ccc",  "#61B329", "#E2B306")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)); figure_2A 


############################################################### WORKS ABOVE HERE AFTER VALIDITY CHECK  

us_pop_bigrams_prc <- bm_text_data %>% 
  unnest_tokens(bigram, abstract, token = "ngrams", n = 2) %>% 
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = black_bigrams), yes = "african-american", no = bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = native_american_bigrams), yes = "native-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = pacific_islander_bigrams), yes = "pacific-islander", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = hispanic_american_bigrams), yes = "hispanic-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = asian_american_bigrams), yes = "asian-american", no = rc_bigram)) %>%
  select(-bigram) %>% 
  filter(rc_bigram == "african-american" | 
         rc_bigram == "native-american" | 
         rc_bigram == "hispanic-american" |
         rc_bigram == "asian-american" |
         rc_bigram == "pacific-islander") %>% 
  rename(term = rc_bigram) %>% 
  select(pubmed_id, year, term) 

us_pop_combined <- bm_abstract_data %>% 
  mutate(term = ifelse(test = str_detect(string = word,
                       pattern = omb_black), yes = "black", no = word)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_white), yes = "white", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_asian), yes = "asian", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                      pattern = omb_hispanic_latinx), yes = "hispanic/latinx", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_racial), yes = "race", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_ethnicity), yes = "ethnicity", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_native_american), yes = "native-american", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_pacific_islander), yes = "pacific-islander", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_diversity), yes = "diversity", no = term)) %>% 
  filter(term == "black" | term == "white" | term == "asian" |
         term == "hispanic/latinx" | term == "race" | term == "ethnicity" |
         term == "native-american" | term == "pacific-islander" | term == "diversity" ) %>% 
  select(pubmed_id, year, term) %>% 
  bind_rows(us_pop_bigrams_prc)

# graphing us-specific terms over time 
us_pop_combined <- us_pop_combined %>% 
  mutate(black = ifelse(test = str_detect(string = term, 
                       pattern = "black"), yes = 1, no = 0)) %>% 
  mutate(white = ifelse(test = str_detect(string = term, 
                       pattern = "white"), yes = 1, no = 0)) %>%
  mutate(asian = ifelse(test = str_detect(string = term, 
                       pattern = "asian"), yes = 1, no = 0)) %>%
  mutate(hispanic = ifelse(test = str_detect(string = term, 
                       pattern = "hispanic/latinx"), yes = 1, no = 0)) %>% 
  mutate(native_american = ifelse(test = str_detect(string = term, 
                       pattern = "native-american"), yes = 1, no = 0)) %>% 
  mutate(pacific_islander = ifelse(test = str_detect(string = term, 
                       pattern = "pacific-islander"), yes = 1, no = 0)) %>% 
  mutate(racial = ifelse(test = str_detect(string = term, 
                       pattern = "race"), yes = 1, no = 0)) %>% 
  mutate(ethnic = ifelse(test = str_detect(string = term, 
                       pattern = "ethnicity"), yes = 1, no = 0)) %>% 
  mutate(diversity = ifelse(test = str_detect(string = term, 
                       pattern = "diversity"), yes = 1, no = 0))

# total articles each year 
us_pop_prc_counts_2B <- bm_text_data %>% 
  drop_na(abstract) %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n)  

# articles with term mentioned each year 
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(black == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_black = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_black = round(cnt_black / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(white == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_white = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_white = round(cnt_white / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(asian == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_asian = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_asian = round(cnt_asian / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(hispanic == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_hispanic = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_hispanic = round(cnt_hispanic / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(native_american == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_native_american = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_native_american = round(cnt_native_american / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(pacific_islander == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_pacific_islander = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_pacific_islander = round(cnt_pacific_islander / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(racial == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_racial = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_racial = round(cnt_racial / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(ethnic == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_ethnic = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_ethnic = round(cnt_ethnic / total * 100, digits = 2))
us_pop_prc_counts_2B <- us_pop_combined %>% 
  filter(diversity == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_diversity = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2B, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2))

# graphing terms over time 
figure_2B <- us_pop_prc_counts_2B %>% ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size = 1.3, stat="identity") +
  labs(title = "    Figure 2B. Proportional Growth of \n US-Specific Terms in Biomedical Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(labels=c("asian","black","diversity","ethnicity","hispanic/latinx",
                              "native-american","pacific-islander","race","white"),
                     values=c("#E86F00", "#3B9AB2", "#000000", "#E21F00", "#00468b", 
                              "#6b0f6b", "#d23ccc",  "#61B329", "#E2B306")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)); figure_2B

legend_2AB <- us_pop_prc_counts_2B %>% ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size = 1.3, stat="identity") +
  labs(title = "    Figure 2B. Proportional Growth of \n US-Specific Terms in Biomedical Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        #plot.title = element_text(size=12, hjust = 0.5, face="bold")
        ) +
  scale_color_manual(labels=c("asian","black","diversity","ethnicity","hispanic/latinx",
                              "native-american","pacific-islander","race","white"),
                     values=c("#E86F00", "#3B9AB2", "#000000", "#E21F00", "#00468b", 
                              "#6b0f6b", "#d23ccc",  "#61B329", "#E2B306")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_2AB <- cowplot::get_legend(legend_2AB)

#grid.arrange(general_pop_graph, gen_pop_prc_counts_graph, nrow = 1, ncol = 2)

grid.arrange(figure_2A, figure_2B, 
             legend_2AB, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )

```

```{r figures_2CD, fig.height=5, fig.width=11.5, message = FALSE, results = FALSE, warning = FALSE, echo=FALSE}
setwd("~/Documents/Diversity/Data")
# loading the .csv file 
div_text_data <- read_csv("diversity_corpus.csv") %>% 
  drop_na(abstract) %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018") 

my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier", 
                                "women.recommendations", 
                                "cs.man.ac.uk", "sexes.please", "movements.baby"))

# tokenizing the abstract data into words 
div_abstract_data <- div_text_data %>% 
  unnest_tokens(word, abstract) %>% 
  anti_join(stop_words) %>% 
  anti_join(my_stopwords); rm(my_stopwords)

black_bigrams <- "\\b(?i)(african american|african americans|afro american|afro americans|african-american|african-americans|afro-american|afro-americans)\\b"
native_american_bigrams <- "\\b(?i)(native american|native americans|american indian|american indians|alaskan native|alaskan natives|alaska nativez|alaska natives|native-american|native-americans|american-indian|american-indians|alaskan-native|alaskan-natives|alaska-native|alaska-natives)\\b"
pacific_islander_bigrams <- "\\b(?i)(pacific islander|pacific islanders|native hawaiian|native hawaiians|pacific-islander|pacific-islanders|native-hawaiian|native-hawaiians)\\b"
hispanic_american_bigrams <- "\\b(?i)(hispanic-american|hispanic american|hispanic-americans|hispanic americans)\\b"
asian_american_bigrams <- "\\b(?i)(asian-american|asian american)\\b"
asian_american_bigrams <- "\\b(?i)(caucasian-american|caucasian american)\\b"

setwd("~/Documents/Diversity/Data")
h1_dictionary <- read_csv("diversity_project - h1_dictionary.csv")
h3_dictionary <- read_csv("diversity_project - h3_dictionary.csv")
omb_black <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$black), "z3x)\\b"), collapse = "|")
omb_native_american <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$native_american), "z3x)\\b"), collapse = "|")
omb_pacific_islander <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$pacific_islander), "z3x)\\b"), collapse = "|")
omb_hispanic_latinx <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$hispanic_latinx), "z3x)\\b"), collapse = "|")
omb_asian <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$asian), "z3x)\\b"), collapse = "|")
omb_white <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$white), "z3x)\\b"), collapse = "|")
omb_racial <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$race), "z3x)\\b"), collapse = "|")
omb_ethnicity <- paste(c("\\b(?i)(z3x",na.omit(h3_dictionary$ethnicity), 
                         "z3x)\\b"), collapse = "|")
omb_diversity <- paste(c("\\b(?i)(z3x",na.omit(h1_dictionary$diversity), 
                         "z3x)\\b"), collapse = "|")

# recoding frequently occuring population-related bigrams 
div_us_pop_bigrams <- div_text_data %>% 
  unnest_tokens(bigram, abstract, token = "ngrams", n = 2) %>% 
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = black_bigrams), yes = "african-american", no = bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = native_american_bigrams), yes = "native-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = pacific_islander_bigrams), yes = "pacific-islander", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = hispanic_american_bigrams), yes = "hispanic-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = asian_american_bigrams), yes = "asian-american", no = rc_bigram)) %>%
  select(-bigram) %>% 
  filter(rc_bigram == "african-american" | 
         rc_bigram == "native-american" | 
         rc_bigram == "hispanic-american" |
         rc_bigram == "asian-american" |
         rc_bigram == "pacific-islander") %>% 
  count(rc_bigram, year, sort = TRUE) %>% 
  rename(word = rc_bigram)

# combining the hyphenated words to our full dataset 
div_us_pop_counts <- div_abstract_data %>% 
  group_by(year) %>% 
  count(word, sort = TRUE) %>% 
  ungroup() %>% 
  select(word, year, n) %>% 
  bind_rows(div_us_pop_bigrams) %>% 
  arrange(-n)

div_us_pop_counts <- div_us_pop_counts %>% 
  mutate(term = ifelse(test = str_detect(string = word,
                       pattern = omb_black), yes = "black", no = word)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_white), yes = "white", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_asian), yes = "asian", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                      pattern = omb_hispanic_latinx), yes = "hispanic/latinx", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_racial), yes = "race", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_ethnicity), yes = "ethnicity", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_native_american), yes = "native-american", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_pacific_islander), yes = "pacific-islander", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_diversity), yes = "diversity", no = term)) %>% 
  filter(term == "black" | term == "white" | term == "asian" |
         term == "hispanic/latinx" | term == "race" | term == "ethnicity" |
         term == "native-american" | term == "pacific-islander" | term == "diversity" ) %>%  
  group_by(term, year) %>% 
  summarise(n = sum(n)) %>% 
  arrange(-n)

# graphing us-specific terms over time using census categories 
# https://www.census.gov/topics/population/race/about.html
figure_2C <- ggplot(div_us_pop_counts) + 
  geom_line(aes(y = n, x = year, colour = term), 
            size = 1.5, stat="identity") + 
  theme_bw() +
  labs(title = "     Figure 2C. Total Growth of \n US-Specific Terms in Diversity Sample (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(labels=c("asian","black","diversity","ethnicity","hispanic/latinx",
                              "native-american","pacific-islander","race","white"),
                     values=c("#E86F00", "#3B9AB2", "#000000", "#E21F00", "#00468b", 
                              "#6b0f6b", "#d23ccc",  "#61B329", "#E2B306")) +
  scale_y_log10() +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)); figure_2C 


############################################################### WORKS ABOVE HERE AFTER VALIDITY CHECK  

div_us_pop_bigrams_prc <- div_text_data %>% 
  unnest_tokens(bigram, abstract, token = "ngrams", n = 2) %>% 
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = black_bigrams), yes = "african-american", no = bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = native_american_bigrams), yes = "native-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = pacific_islander_bigrams), yes = "pacific-islander", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = hispanic_american_bigrams), yes = "hispanic-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
         pattern = asian_american_bigrams), yes = "asian-american", no = rc_bigram)) %>%
  select(-bigram) %>% 
  filter(rc_bigram == "african-american" | 
         rc_bigram == "native-american" | 
         rc_bigram == "hispanic-american" |
         rc_bigram == "asian-american" |
         rc_bigram == "pacific-islander") %>% 
  rename(term = rc_bigram) %>% 
  select(pubmed_id, year, term) 

div_us_pop_combined <- div_abstract_data %>% 
  mutate(term = ifelse(test = str_detect(string = word,
                       pattern = omb_black), yes = "black", no = word)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_white), yes = "white", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_asian), yes = "asian", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                      pattern = omb_hispanic_latinx), yes = "hispanic/latinx", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_racial), yes = "race", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_ethnicity), yes = "ethnicity", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_native_american), yes = "native-american", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_pacific_islander), yes = "pacific-islander", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = omb_diversity), yes = "diversity", no = term)) %>% 
  filter(term == "black" | term == "white" | term == "asian" |
         term == "hispanic/latinx" | term == "race" | term == "ethnicity" |
         term == "native-american" | term == "pacific-islander" | term == "diversity" ) %>% 
  select(pubmed_id, year, term) %>% 
  bind_rows(div_us_pop_bigrams_prc)

# graphing us-specific terms over time 
div_us_pop_combined <- div_us_pop_combined %>% 
  mutate(black = ifelse(test = str_detect(string = term, 
                       pattern = "black"), yes = 1, no = 0)) %>% 
  mutate(white = ifelse(test = str_detect(string = term, 
                       pattern = "white"), yes = 1, no = 0)) %>%
  mutate(asian = ifelse(test = str_detect(string = term, 
                       pattern = "asian"), yes = 1, no = 0)) %>%
  mutate(hispanic = ifelse(test = str_detect(string = term, 
                       pattern = "hispanic/latinx"), yes = 1, no = 0)) %>% 
  mutate(native_american = ifelse(test = str_detect(string = term, 
                       pattern = "native-american"), yes = 1, no = 0)) %>% 
  mutate(pacific_islander = ifelse(test = str_detect(string = term, 
                       pattern = "pacific-islander"), yes = 1, no = 0)) %>% 
  mutate(racial = ifelse(test = str_detect(string = term, 
                       pattern = "race"), yes = 1, no = 0)) %>% 
  mutate(ethnic = ifelse(test = str_detect(string = term, 
                       pattern = "ethnicity"), yes = 1, no = 0)) %>% 
  mutate(diversity = ifelse(test = str_detect(string = term, 
                       pattern = "diversity"), yes = 1, no = 0))

# total articles each year 
us_pop_prc_counts_2D <- div_text_data %>% 
  drop_na(abstract) %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n)  

# articles with term mentioned each year 
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(black == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_black = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_black = round(cnt_black / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(white == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_white = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_white = round(cnt_white / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(asian == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_asian = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_asian = round(cnt_asian / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(hispanic == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_hispanic = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_hispanic = round(cnt_hispanic / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(native_american == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_native_american = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_native_american = round(cnt_native_american / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(pacific_islander == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_pacific_islander = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_pacific_islander = round(cnt_pacific_islander / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(racial == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_racial = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_racial = round(cnt_racial / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(ethnic == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_ethnic = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_ethnic = round(cnt_ethnic / total * 100, digits = 2))
us_pop_prc_counts_2D <- div_us_pop_combined %>% 
  filter(diversity == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_diversity = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts_2D, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2))

# graphing terms over time 
figure_2D <- us_pop_prc_counts_2D %>% ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size = 1.3, stat="identity") +
  labs(title = "    Figure 2D. Proportional Growth of \n US-Specific Terms in Diversity Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(labels=c("asian","black","diversity","ethnicity","hispanic/latinx",
                              "native-american","pacific-islander","race","white"),
                     values=c("#E86F00", "#3B9AB2", "#000000", "#E21F00", "#00468b", 
                              "#6b0f6b", "#d23ccc",  "#61B329", "#E2B306")) +
  scale_y_log10() +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)); figure_2B

legend_2CD <- us_pop_prc_counts_2D %>% ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 1.3, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 1.3,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 1.3, stat="identity") +
  geom_line(aes(y = prc_diversity, x = year, colour = "diversity"), size = 1.3, stat="identity") +
  labs(title = "    Figure 2D. Proportional Growth of \n US-Specific Terms in Diversity Sample (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Abstracts") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        #plot.title = element_text(size=12, hjust = 0.5, face="bold")
        ) +
  scale_color_manual(labels=c("asian","black","diversity","ethnicity","hispanic/latinx",
                              "native-american","pacific-islander","race","white"),
                     values=c("#E86F00", "#3B9AB2", "#000000", "#E21F00", "#00468b", 
                              "#6b0f6b", "#d23ccc",  "#61B329", "#E2B306")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_2CD <- cowplot::get_legend(legend_2CD)

#grid.arrange(general_pop_graph, gen_pop_prc_counts_graph, nrow = 1, ncol = 2)

grid.arrange(figure_2C, figure_2D, 
             legend_2CD, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )

```



```{r binding 3ab + 3cd, fig.height=7, fig.width=11.5, echo=FALSE}
lay <- rbind(c(1,2),
             c(3,4),
             c(5,5))

grid.arrange(figure_2A, figure_2B, 
             figure_2C, figure_2D,
             legend_2AB, 
             nrow = 3, ncol=2,
             layout_matrix = lay,
             #widths = c(2.7, 2.7), 
             heights = c(4.2, 4.2, 0.8)
             )
```




HEREHHEHREHERHEHHERE

```{r figures 2ab, fig.height=5, fig.width=11.5, echo=FALSE}
for (pkg in c("tidyverse", "igraph", "tidytext", "ggplot2", "DT", "maditr",
              "grid", "gridExtra", "reshape2", "extrafont", "data.table",
              "plotly", "stringi", "networkD3")) {library(pkg, character.only = TRUE)}

setwd("~/Documents/Diversity/Data")
# loading the .csv file 
text_data <- read_csv("biomedical_corpus.csv") %>% 
  rowid_to_column(var = "id") %>% 
  rename(authors = AU, title = TI, publication = SO, #author_keywords = DE,
         abstract = AB, references = CR, year = PY, times_cited = TC, pubmed_id = UT) %>% 
  select(id, authors, title, year, publication, abstract, references, year, times_cited, pubmed_id) %>% 
  drop_na(abstract) %>%
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018")

# tokenizing the abstract data into words 
abstract_data <- text_data %>% 
  unnest_tokens(word, abstract) #%>% anti_join(stop_words)
# adding custom set of stopwords 
my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier", 
                                "women.recommendations", 
                                "cs.man.ac.uk", "sexes.please", "movements.baby"))
abstract_data <- abstract_data %>% 
  anti_join(my_stopwords)

# lets draw all of our strings from the diversity_dictionary
setwd("~/Documents/Diversity/Data")
divictionary <- read_csv("diversity_project - h1_dictionary.csv") 

# pulling our list of population
h2_dictionary <- read_csv("diversity_project - h2_dictionary.csv") 
# pulling in concatenated population terms from our dataset 
continental_terms <- h2_dictionary %>% filter(category == "continental")
subcontinental_terms <- h2_dictionary %>% filter(category == "subcontinental")
national_terms <- h2_dictionary %>% filter(category == "national")
subnational_terms <- h2_dictionary %>% filter(category == "subnational")
directional_terms <- h2_dictionary %>% filter(category == "directional")
race_ethnicity <- h2_dictionary %>% filter(category == "race/ethnicity")
omb_uscensus <- h2_dictionary %>% filter(category == "omb/us census")
ancestry <- h2_dictionary %>% filter(category == "ancestry")

# creating detection string patterns for all the categories
all_pop_terms <- paste(c("\\b(?i)(zcx", 
                 h2_dictionary$term, "zxc)\\b"), collapse = "|")
continental_terms <- paste(c("\\b(?i)(zxz", 
                     continental_terms$term, "zxz)\\b"), collapse = "|")
subcontinental_terms <- paste(c("\\b(?i)(zxz", 
                        subcontinental_terms$term, "zxz)\\b"), collapse = "|")
national_terms <- paste(c("\\b(?i)(zxz", 
                  national_terms$term, "zxz)\\b"), collapse = "|")
subnational_terms <- paste(c("\\b(?i)(zxz", 
                     subnational_terms$term, "zxz)\\b"), collapse = "|")
directional_terms <- paste(c("\\b(?i)(zxz", 
                     directional_terms$term, "zxz)\\b"), collapse = "|")
race_ethnicity <- paste(c("\\b(?i)(zxz", 
                  race_ethnicity$term, "zxz)\\b"), collapse = "|")
omb_uscensus <- paste(c("\\b(?i)(zxz", 
                omb_uscensus$term, "zxz)\\b"), collapse = "|")
ancestry <- paste(c("\\b(?i)(zxz", 
            ancestry$term, "zxz)\\b"), collapse = "|")

# graphing the population terms into various categories 
pop_terms_bycats <- abstract_data %>% 
  as.data.table() %>% 
  dt_mutate(all_pop_terms = ifelse(test = str_detect(string = abstract_data$word, 
                         pattern = all_pop_terms), 
                         yes = 1, no = 0)) %>%
  dt_mutate(continental = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = continental_terms), 
                       yes = 1, no = 0)) %>%
  dt_mutate(subcontinental = ifelse(test = str_detect(string = abstract_data$word, 
                          pattern = subcontinental_terms), 
                          yes = 1, no = 0)) %>%
  dt_mutate(national = ifelse(test = str_detect(string = abstract_data$word, 
                    pattern = national_terms), 
                    yes = 1, no = 0)) %>%
  dt_mutate(subnational = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = subnational_terms), 
                       yes = 1, no = 0)) %>%
  dt_mutate(directional = ifelse(test = str_detect(string = abstract_data$word, 
                       pattern = directional_terms), 
                       yes = 1, no = 0)) %>%
  dt_mutate(race_ethnicity = ifelse(test = str_detect(string = abstract_data$word, 
                          pattern = race_ethnicity), 
                          yes = 1, no = 0)) %>%
  dt_mutate(omb_uscensus = ifelse(test = str_detect(string = abstract_data$word, 
                          pattern = omb_uscensus), 
                          yes = 1, no = 0)) %>%
  dt_mutate(ancestry = ifelse(test = str_detect(string = abstract_data$word, 
                          pattern = ancestry), 
                          yes = 1, no = 0))

h2_counts <- pop_terms_bycats %>% 
  filter(all_pop_terms == 1) %>%
  group_by(year) %>% 
  count(all_pop_terms) %>% 
  arrange(year) %>% 
  mutate(term = "all population terms") %>% 
  select(term, n, year) 

h2_counts <- pop_terms_bycats %>% 
  filter(continental == 1) %>%
  group_by(year) %>% 
  count(continental) %>% 
  arrange(year) %>% 
  mutate(term = "continental") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(subcontinental == 1) %>%
  group_by(year) %>% 
  count(subcontinental) %>% 
  arrange(year) %>% 
  mutate(term = "subcontinental") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(national == 1) %>%
  group_by(year) %>% 
  count(national) %>% 
  arrange(year) %>% 
  mutate(term = "national") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(subnational == 1) %>%
  group_by(year) %>% 
  count(subnational) %>% 
  arrange(year) %>% 
  mutate(term = "subnational") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(directional == 1) %>%
  group_by(year) %>% 
  count(directional) %>% 
  arrange(year) %>% 
  mutate(term = "directional") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(race_ethnicity == 1) %>%
  group_by(year) %>% 
  count(race_ethnicity) %>% 
  arrange(year) %>% 
  mutate(term = "race/ethnicity") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(omb_uscensus == 1) %>%
  group_by(year) %>% 
  count(omb_uscensus) %>% 
  arrange(year) %>% 
  mutate(term = "omb/us-census") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

h2_counts <- pop_terms_bycats %>% 
  filter(ancestry == 1) %>%
  group_by(year) %>% 
  count(ancestry) %>% 
  arrange(year) %>% 
  mutate(term = "ancestry") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_counts)

# graphing the patterns over time
figure_3A <- h2_counts %>%   
  #filter(term != "all population terms") %>% 
  ggplot() + geom_line(aes(y = n, x = year, colour = term), size = 1.3, stat="identity") + 
  labs(title = "     Figure 3A. Total Growth of Population Terms \n in Biomedical Corpus (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "ancestry", "continental", 
                              "directional", "national", "omb/us-census", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00","#6b0f6b", "#f2be00", 
                              "#00468b", "#E86F00","#61B329", 
                              "#d23ccc", "#3B9AB2", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

## RUN HERE

# total articles each year 
bycats_prc_counts <- text_data %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n) 

# articles with term mentioned each year 
all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(all_pop_terms == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "all population terms") %>% 
  select(term, n, year)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(continental == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "continental") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(subcontinental == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "subcontinental") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(national == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "national") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(subnational == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "subnational") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(directional == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "directional") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(race_ethnicity == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "race/ethnicity") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(omb_uscensus == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "omb/us-census") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- pop_terms_bycats %>% 
  filter(ancestry == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "ancestry") %>% 
  select(term, n, year) %>% 
  bind_rows(all_pop_prc_counts)

all_pop_prc_counts <- all_pop_prc_counts %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(percentage = round(n / total * 100, digits = 2))


# legend plot 
figure_3B <- all_pop_prc_counts %>% 
  ggplot() + 
  geom_line(aes(y = percentage, x = year, colour = term), size = 1.3, stat="identity") +
  labs(title = "    Figure 3B. Proportional Growth of Population Terms \n in Biomedical Corpus (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "ancestry", "continental", 
                              "directional", "national", "omb/us-census", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00","#6b0f6b", "#f2be00", 
                              "#00468b", "#E86F00","#61B329", 
                              "#d23ccc", "#3B9AB2", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

figure_3AB <- all_pop_prc_counts %>% 
  ggplot() + 
  geom_line(aes(y = percentage, x = year, colour = term), size = 1.3, stat="identity") +
  labs(title = "    Figure 3B. Proportional Growth of Population Terms \n in Biomedical Corpus (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        #plot.title = element_text(size=12, hjust = 0.5, face="bold")
        ) +
  scale_color_manual(labels=c("all population terms", "ancestry", "continental", 
                              "directional", "national", "omb/us-census", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00","#6b0f6b", "#f2be00", 
                              "#00468b", "#E86F00","#61B329", 
                              "#d23ccc", "#3B9AB2", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

figure_3AB <- cowplot::get_legend(figure_3AB)

#grid.arrange(general_pop_graph, gen_pop_prc_counts_graph, nrow = 1, ncol = 2)

grid.arrange(figure_3A, figure_3B, 
             figure_3AB, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )
#ggplotly(figure_2B)
```


```{r figures 3cd, fig.height=5, fig.width=11.5, echo=FALSE}

setwd("~/Documents/Diversity/Data")
# loading the .csv file 
div_text_data <- read_csv("diversity_corpus.csv") %>% 
  drop_na(abstract) %>% 
# filtering 2018 articles because they seem to be incomplete 
  filter(year != "2018") 
# tokenizing the abstract data into words 
abstract_div_data <- div_text_data %>% 
  unnest_tokens(word, abstract) %>% 
  anti_join(stop_words)
# adding custom set of stopwords 
my_stopwords <- tibble(word = c(as.character(1:14), "[0-9]", 
                                "rights", "reserved", "copyright", "elsevier", 
                                "women.recommendations", 
                                "cs.man.ac.uk", "sexes.please", "movements.baby"))
abstract_div_data <- abstract_div_data %>% 
  anti_join(my_stopwords)

# graphing the population terms into various categories 
diversity_bycats <- abstract_div_data %>% 
  as.data.table() %>% 
  dt_mutate(all_pop_terms = ifelse(test = str_detect(string = word, 
                         pattern = all_pop_terms), 
                         yes = 1, no = 0)) %>%
  dt_mutate(continental = ifelse(test = str_detect(string = word, 
                       pattern = continental_terms), 
                       yes = 1, no = 0)) %>%
  dt_mutate(subcontinental = ifelse(test = str_detect(string = word, 
                          pattern = subcontinental_terms), 
                          yes = 1, no = 0)) %>%
  dt_mutate(national = ifelse(test = str_detect(string = word, 
                    pattern = national_terms), 
                    yes = 1, no = 0)) %>%
  dt_mutate(subnational = ifelse(test = str_detect(string = word, 
                       pattern = subnational_terms), 
                       yes = 1, no = 0)) %>%
  dt_mutate(directional = ifelse(test = str_detect(string = word, 
                       pattern = directional_terms), 
                       yes = 1, no = 0)) %>%
  dt_mutate(race_ethnicity = ifelse(test = str_detect(string = word, 
                          pattern = race_ethnicity), 
                          yes = 1, no = 0)) %>%
  dt_mutate(omb_uscensus = ifelse(test = str_detect(string = word, 
                          pattern = omb_uscensus), 
                          yes = 1, no = 0)) %>%
  dt_mutate(ancestry = ifelse(test = str_detect(string = word, 
                          pattern = ancestry), 
                          yes = 1, no = 0))

h2_div_counts <- diversity_bycats %>% 
  filter(all_pop_terms == 1) %>%
  group_by(year) %>% 
  count(all_pop_terms) %>% 
  arrange(year) %>% 
  mutate(term = "all population terms") %>% 
  select(term, n, year) 

h2_div_counts <- diversity_bycats %>% 
  filter(continental == 1) %>%
  group_by(year) %>% 
  count(continental) %>% 
  arrange(year) %>% 
  mutate(term = "continental") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(subcontinental == 1) %>%
  group_by(year) %>% 
  count(subcontinental) %>% 
  arrange(year) %>% 
  mutate(term = "subcontinental") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(national == 1) %>%
  group_by(year) %>% 
  count(national) %>% 
  arrange(year) %>% 
  mutate(term = "national") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(subnational == 1) %>%
  group_by(year) %>% 
  count(subnational) %>% 
  arrange(year) %>% 
  mutate(term = "subnational") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(directional == 1) %>%
  group_by(year) %>% 
  count(directional) %>% 
  arrange(year) %>% 
  mutate(term = "directional") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(race_ethnicity == 1) %>%
  group_by(year) %>% 
  count(race_ethnicity) %>% 
  arrange(year) %>% 
  mutate(term = "race/ethnicity") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(omb_uscensus == 1) %>%
  group_by(year) %>% 
  count(omb_uscensus) %>% 
  arrange(year) %>% 
  mutate(term = "omb/us-census") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

h2_div_counts <- diversity_bycats %>% 
  filter(ancestry == 1) %>%
  group_by(year) %>% 
  count(ancestry) %>% 
  arrange(year) %>% 
  mutate(term = "ancestry") %>% 
  select(term, n, year) %>% 
  bind_rows(h2_div_counts)

figure_3C <- h2_div_counts %>%   
  #filter(term != "all population terms") %>% 
  ggplot() + geom_line(aes(y = n, x = year, colour = term), size = 1.3, stat="identity") + 
  theme_bw() +
  labs(title = "     Figure 3C. Total Growth of Population Terms \n in Diversity Corpus (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "ancestry", "continental", 
                              "directional", "national", "omb/us-census", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00","#6b0f6b", "#f2be00", 
                              "#00468b", "#E86F00","#61B329", 
                              "#d23ccc", "#3B9AB2", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

# total articles each year 
div_prc_counts <- div_text_data %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n) 

# articles with term mentioned each year 
div_prc_counts <- diversity_bycats %>% 
  filter(all_pop_terms == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "all population terms") %>% 
  select(term, n, year)

div_prc_counts <- diversity_bycats %>% 
  filter(continental == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "continental") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(subcontinental == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "subcontinental") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(national == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "national") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(subnational == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "subnational") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(directional == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "directional") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(race_ethnicity == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "race/ethnicity") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(omb_uscensus == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "omb/us-census") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- diversity_bycats %>% 
  filter(ancestry == 1) %>% 
  group_by(year) %>% 
  summarise(n = n_distinct(pubmed_id)) %>% 
  mutate(term = "ancestry") %>% 
  select(term, n, year) %>% 
  bind_rows(div_prc_counts)

div_prc_counts <- div_prc_counts %>% 
  right_join(bycats_prc_counts, by = "year") %>% 
  mutate(percentage = round(n / total * 100, digits = 2))

figure_3D <- div_prc_counts %>% 
  ggplot() + 
  geom_line(aes(y = percentage, x = year, colour = term), size = 1.3, stat="identity") +
  labs(title = "    Figure 3D. Proportional Growth of Population Terms \n in Diversity Corpus (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5)) +
  scale_color_manual(labels=c("all population terms", "ancestry", "continental", 
                              "directional", "national", "omb/us-census", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00","#6b0f6b", "#f2be00", 
                              "#00468b", "#E86F00","#61B329", 
                              "#d23ccc", "#3B9AB2", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

#legend 

legend_3CD <- div_prc_counts %>% 
  ggplot() + 
  geom_line(aes(y = percentage, x = year, colour = term), size = 1.3, stat="identity") +
  labs(title = "    Figure 3D. Proportional Growth of Population Terms \n in Diversity Corpus (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(labels=c("all population terms", "ancestry", "continental", 
                              "directional", "national", "omb/us-census", 
                              "race/ethnicity", "subcontinental", "subnational"),
                     values=c("#E21F00","#6b0f6b", "#f2be00", 
                              "#00468b", "#E86F00","#61B329", 
                              "#d23ccc", "#3B9AB2", "#1b7524")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_3CD <- cowplot::get_legend(legend_3CD)

#grid.arrange(general_pop_graph, gen_pop_prc_counts_graph, nrow = 1, ncol = 2)

grid.arrange(figure_3C, figure_3D, 
             legend_3CD, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )
```

```{r binding 3ab + 3cd, fig.height=7, fig.width=11.5, echo=FALSE}
lay <- rbind(c(1,2),
             c(3,4),
             c(5,5))

grid.arrange(figure_3A, figure_3B, 
             figure_3C, figure_3D,
             legend_3CD, 
             nrow = 3, ncol=2,
             layout_matrix = lay,
             #widths = c(2.7, 2.7), 
             heights = c(4.2, 4.2, 0.8)
             )
```

UPDATED BEFORE THIS 


```{r figure 2gh, fig.height=5, fig.width=11.5, echo=FALSE}
# recoding frequently occuring population-related bigrams 
us_pop_bigrams <- text_data %>% 
  unnest_tokens(bigram, abstract, token = "ngrams", n = 2) %>% 
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(african american|african americans|afro american|afro americans|african-american|african-americans|afro-american|afro-americans)\\b"), 
                       yes = "african-american", no = bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(native american|native americans|american indian|american indians|alaskan native|alaskan natives|alaska nativez|alaska natives|native-american|native-americans|american-indian|american-indians|alaskan-native|alaskan-natives|alaska-native|alaska-natives)\\b"), yes = "native-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(pacific islander|pacific islanders|native hawaiian|native hawaiians|pacific-islander|pacific-islanders|native-hawaiian|native-hawaiians)\\b"), 
                       yes = "pacific-islander", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(mexican american|mexican americans|mexican-american|mexican-americans|hispanic-american|hispanic american|hispanic-americans|hispanic americans)\\b"), 
                       yes = "hispanic-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(asian-american|asian american|cambodian american|cambodian americans|chinese american|chinese americans|indian american|indian americans|japanese american|japanese americans|korean american|korean americans|malaysian american|malaysian americans|pakistani american|pakistani americans|filipino american|filipino americans|thai american|thai americans|vietnamese american|vietnamese americans|cambodian-american|cambodian-americans|chinese-american|chinese-americans|indian-american|indian-americans|japanese-american|japanese-americans|korean-american|korean-americans|malaysian-american|malaysian-americans|pakistani-american|pakistani-americans|filipino-american|filipino-americans|thai-american|thai-americans|vietnamese-american|vietnamese-americans)\\b"), 
  yes = "asian-american", no = rc_bigram)) %>%
  select(-bigram) %>% 
  filter(rc_bigram == "african-american" | 
         rc_bigram == "native-american" | 
         rc_bigram == "hispanic-american" |
         rc_bigram == "asian-american" |
         rc_bigram == "pacific-islander") %>% 
  count(rc_bigram, year, sort = TRUE) %>% 
  rename(word = rc_bigram)

# combining the hyphenated words to our full dataset 
us_pop_counts <- abstract_data %>% 
  group_by(year) %>% 
  count(word, sort = TRUE) %>% 
  ungroup() %>% 
  select(word, year, n) %>% 
  bind_rows(us_pop_bigrams) %>% 
  arrange(-n)

# graphing us-specific terms over time using census categories 
# https://www.census.gov/topics/population/race/about.html
figure_4C <- us_pop_counts %>% 
  mutate(term = ifelse(test = str_detect(string = word,
                       pattern = "\\b(?i)(black|blacks|negro|negros|african-american)\\b"), 
                       yes = "black", no = word)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(white|whites|caucasian|caucasians)\\b"), 
                       yes = "white", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(asian|asians|asian-american)\\b"), 
                       yes = "asian", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
  pattern = "\\b(?i)(hispanic|hispanic-american|latino|latinos|latina|latinos|latinx|latinxs|latino/a|latino/as)\\b"), 
                       yes = "hispanic/latinx", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(race|racial|racially)\\b"), 
                       yes = "race", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(ethnic|ethnicity|ethnically)\\b"), 
                       yes = "ethnicity", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(native-american)\\b"), 
                       yes = "native-american", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(pacific-islander)\\b"), 
                       yes = "pacific-islander", no = term)) %>%
  filter(term == "black" | term == "white" | term == "asian" |
         term == "hispanic/latinx" | term == "race" | term == "ethnicity" |
         term == "native-american" | term == "pacific-islander" ) %>%  
  group_by(term, year) %>% summarise(n = sum(n)) %>% arrange(-n) %>% 
  ggplot() + geom_line(aes(y = n, x = year, colour = term), size = 1.5, stat="identity") + 
  theme_bw() +
  labs(title = "     Figure 4C. Total Growth of \n US-Specific Terms (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Term Count") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#61B329", "#3B9AB2", "#E21F00", "#00468b", "#d23ccc", "#6b0f6b",  "#E86F00", "#E2B306")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015)) 
  #scale_linetype_identity( linetype=c("solid", "solid",  "solid", "twodash", "dotted", "dashed")) + 
  #scale_color_manual( values=c( "black", "#666666", "grey","black", "#666666", "grey"))

# recoding frequently occuring population-related bigrams 
us_pop_bigrams <- text_data %>% 
  unnest_tokens(bigram, abstract, token = "ngrams", n = 2) %>% 
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(african american|african americans|afro american|afro americans|african-american|african-americans|afro-american|afro-americans)\\b"), 
                       yes = "african-american", no = bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(native american|native americans|american indian|american indians|alaskan native|alaskan natives|alaska nativez|alaska natives|native-american|native-americans|american-indian|american-indians|alaskan-native|alaskan-natives|alaska-native|alaska-natives)\\b"), yes = "native-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(pacific islander|pacific islanders|native hawaiian|native hawaiians|pacific-islander|pacific-islanders|native-hawaiian|native-hawaiians)\\b"), 
                       yes = "pacific-islander", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(mexican american|mexican americans|mexican-american|mexican-americans|hispanic-american|hispanic american|hispanic-americans|hispanic americans)\\b"), 
                       yes = "hispanic-american", no = rc_bigram)) %>%
  mutate(rc_bigram = ifelse(test = str_detect(string = bigram,
  pattern = "\\b(?i)(asian-american|asian american|cambodian american|cambodian americans|chinese american|chinese americans|indian american|indian americans|japanese american|japanese americans|korean american|korean americans|malaysian american|malaysian americans|pakistani american|pakistani americans|filipino american|filipino americans|thai american|thai americans|vietnamese american|vietnamese americans|cambodian-american|cambodian-americans|chinese-american|chinese-americans|indian-american|indian-americans|japanese-american|japanese-americans|korean-american|korean-americans|malaysian-american|malaysian-americans|pakistani-american|pakistani-americans|filipino-american|filipino-americans|thai-american|thai-americans|vietnamese-american|vietnamese-americans)\\b"), 
  yes = "asian-american", no = rc_bigram)) %>%
  select(-bigram) %>% 
  filter(rc_bigram == "african-american" | 
         rc_bigram == "native-american" | 
         rc_bigram == "hispanic-american" |
         rc_bigram == "asian-american" |
         rc_bigram == "pacific-islander") %>% 
  count(rc_bigram, year, sort = TRUE) %>% 
  rename(word = rc_bigram)

# combining the hyphenated words to our full dataset 
us_pop_counts <- abstract_data %>% 
  group_by(year) %>% 
  count(word, pubmed_id, sort = TRUE) %>% 
  ungroup() %>% 
  select(word, pubmed_id, year, n) %>% 
  bind_rows(us_pop_bigrams) 

# graphing us-specific terms over time 
us_pop_counts <- us_pop_counts %>% 
  mutate(term = ifelse(test = str_detect(string = word,
                       pattern = "\\b(?i)(black|blacks|negro|negros|african-american)\\b"), 
                       yes = "black", no = word)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(white|whites|caucasian|caucasians)\\b"), 
                       yes = "white", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(asian|asians|asian-american)\\b"), 
                       yes = "asian", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
  pattern = "\\b(?i)(hispanic|hispanic-american|latino|latinos|latina|latinos|latinx|latinxs|latino/a|latino/as)\\b"), 
                       yes = "hispanic/latinx", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(race|racial|racially)\\b"), 
                       yes = "race", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(ethnic|ethnicity|ethnically)\\b"), 
                       yes = "ethnicity", no = term)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(native-american)\\b"), 
                       yes = "native-american", no = term)) %>%
  mutate(term = ifelse(test = str_detect(string = word, 
                       pattern = "\\b(?i)(pacific-islander)\\b"), 
                       yes = "pacific-islander", no = term)) %>% 
  
  mutate(black = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(black)\\b"), yes = 1, no = 0)) %>% 
  mutate(white = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(white)\\b"), yes = 1, no = 0)) %>%
  mutate(asian = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(asian)\\b"), yes = 1, no = 0)) %>%
  mutate(hispanic = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(hispanic/latinx)\\b"), yes = 1, no = 0)) %>% 
  mutate(native_american = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(native-american)\\b"), yes = 1, no = 0)) %>% 
  mutate(pacific_islander = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(pacific-islander)\\b"), yes = 1, no = 0)) %>% 
  mutate(racial = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(race)\\b"), yes = 1, no = 0)) %>% 
  mutate(ethnic = ifelse(test = str_detect(string = term, 
                       pattern = "\\b(ethnicity)\\b"), yes = 1, no = 0)) 

# total articles each year 
us_pop_prc_counts <- text_data %>% 
  group_by(year) %>% 
  count(year) %>% 
  ungroup() %>% 
  rename(total = n)  

# articles with term mentioned each year 
us_pop_prc_counts <- us_pop_counts %>% 
  filter(black == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_black = n_distinct(pubmed_id)) %>% 
  left_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_black = round(cnt_black / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% 
  filter(white == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_white = n_distinct(pubmed_id)) %>% 
  left_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_white = round(cnt_white / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% 
  filter(asian == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_asian = n_distinct(pubmed_id)) %>% 
  left_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_asian = round(cnt_asian / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% 
  filter(hispanic == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_hispanic = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_hispanic = round(cnt_hispanic / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% 
  filter(native_american == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_native_american = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_native_american = round(cnt_native_american / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% 
  filter(pacific_islander == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_pacific_islander = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_pacific_islander = round(cnt_pacific_islander / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% 
  filter(racial == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_racial = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_racial = round(cnt_racial / total * 100, digits = 2))
us_pop_prc_counts <- us_pop_counts %>% filter(ethnic == 1) %>% 
  group_by(year) %>% 
  summarise(cnt_ethnic = n_distinct(pubmed_id)) %>% 
  right_join(us_pop_prc_counts, by = "year") %>% 
  mutate(prc_ethnic = round(cnt_ethnic / total * 100, digits = 2))


# graphing terms over time 
figure_4D <- us_pop_prc_counts %>% ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 1.5, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 1.5, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 1.5, stat="identity") +
  labs(title = "    Figure 4D. Proportional Growth of \n US-Specific Terms (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#61B329", "#3B9AB2", "#E21F00", "#00468b", "#d23ccc", "#6b0f6b",  "#E86F00", "#E2B306")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_4CD <- us_pop_prc_counts %>% ggplot() + 
  geom_line(aes(y = prc_black, x = year, colour = "black"), size = 1.5, stat="identity") +
  geom_line(aes(y = prc_white, x = year, colour = "white"),size = 1.5, stat="identity") +
  geom_line(aes(y = prc_asian, x = year, colour = "asian"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_hispanic, x = year, colour = "hispanic/latinx"),size = 1.5, stat="identity") +
  geom_line(aes(y = prc_native_american, x = year, colour = "native-american"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_pacific_islander, x = year, colour = "pacific-islander"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_ethnic, x = year, colour = "ethnicity"), size = 1.5,stat="identity") +
  geom_line(aes(y = prc_racial, x = year, colour = "race"), size = 1.5, stat="identity") +
  labs(title = "    Figure 4D. Proportional Growth of \n US-Specific Terms (1990-2017)",
       color =  "Term") + 
  theme_bw() +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        #legend.position = "none",
        legend.box = "vertical", 
        legend.position = "bottom",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#61B329", "#3B9AB2", "#E21F00", "#00468b", "#d23ccc", "#6b0f6b",  "#E86F00", "#E2B306")) + 
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_4CD <- cowplot::get_legend(legend_4CD)

#grid.arrange(general_pop_graph, gen_pop_prc_counts_graph, nrow = 1, ncol = 2)

grid.arrange(figure_4C, figure_4D, 
             legend_4CD, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )
```

```{r, fig.height=16, fig.width=11.5, echo=FALSE}
lay <- rbind(c(1,2),
             c(3,3),
             c(4,5),
             c(6,6),
             c(7,8),
             c(9,9),
             c(10,11),
             c(12,12)
             )

grid.arrange(figure_2A, figure_2B, 
             legend_2AB,
             figure_3A, figure_3B, 
             legend_3CD,
             figure_4A, figure_4B, 
             legend_4AB,
             figure_4C, figure_4D, 
             legend_4CD,
             nrow = 8, ncol=2,
             layout_matrix = lay,
             #widths = c(2.7, 2.7), 
             heights = c(4, 0.4, 4, 0.7, 4, 0.4, 4, 0.7)
             )
```

```{r figure 3ab, fig.height=5, fig.width=11.5, echo=FALSE}

# loading the recoded subject data 
genetics_heredity <- read_csv("subject_categories.csv") %>% 
  filter(collapsed_subject == "genetics_heredity")
biomedical_studies <- read_csv("subject_categories.csv") %>% 
  filter(collapsed_subject == "biomedical_studies")
social_behavioral <- read_csv("subject_categories.csv") %>% 
  filter(collapsed_subject == "social_behavioral")
other_studies <- read_csv("subject_categories.csv") %>% 
  filter(collapsed_subject == "other")
# recoding the data 
genetics_heredity <- paste(c("\\b(?i)(zxz", genetics_heredity$original_subject, "zxz)\\b"),
                           collapse = "|")
biomedical_studies <- paste(c("\\b(?i)(zxz", biomedical_studies$original_subject, "zxz)\\b"),
                            collapse = "|")
social_behavioral <- paste(c("\\b(?i)(zxz", social_behavioral$original_subject, "zxz)\\b"),
                           collapse = "|")
other_studies <- paste(c("\\b(?i)(zxz", other_studies$original_subject, "zxz)\\b"), 
                       collapse = "|")

# recoding the subject categories and graphing changes over time 
pop_terms_bysubject_pt1 <- pop_terms_bysubject %>% 
  mutate(category = ifelse(test = str_detect(string = subject, 
                           pattern = genetics_heredity), 
                           yes = "genetics & heredity", no = "")) %>% 
  mutate(category = ifelse(test = str_detect(string = subject, 
                           pattern = biomedical_studies), 
                           yes = "biomedical studies", no = category)) %>% 
  mutate(category = ifelse(test = str_detect(string = subject, 
                           pattern = social_behavioral), 
                           yes = "social & behavioral sciences", no = category)) %>%
  mutate(category = ifelse(test = str_detect(string = subject, 
                           pattern = other_studies), 
                           yes = "other subjects", no = category)) 
# graphing trends 
figure_3A <- pop_terms_bysubject_pt1 %>%
  group_by(year) %>% 
  count(term, category, sort = TRUE) %>% 
  ungroup() %>% 
  filter(term == "all population terms") %>% 
  ggplot() + geom_line(aes(y = n, x = year, colour = category), size = 1.5, stat="identity") +
  theme_bw() +
  labs(title = "     Figure 3A. Total Growth of Population \n Terms by Collapsed Academic Field (1990-2017)",
       color =  "Field") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Number of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#E21F00", "#61B329", "#E2B306", "#3B9AB2")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

# total articles each year 
bysbjrc_prc_counts <- text_data %>% group_by(year) %>% 
  count(year) %>% ungroup() %>% rename(total = n) 
# articles with term mentioned each year 
bysbjrc_prc_counts <- pop_terms_bysubject_pt1 %>% filter(diversity == 1) %>% 
  group_by(category, year) %>% summarise(cnt_diversity = n_distinct(pubmed_id)) %>% 
  right_join(bysbjrc_prc_counts, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2))
# wrangling all cats into a df 
bysbjrc_graph <- bysbjrc_prc_counts %>% filter(category == "genetics & heredity") %>% 
  mutate(genetics_heredity = prc_diversity) 
bysbjrc_graph <- bysbjrc_prc_counts %>% filter(category == "biomedical studies") %>% 
  mutate(biomedical_studies = prc_diversity) %>% right_join(bysbjrc_graph, by = "year")
bysbjrc_graph <- bysbjrc_prc_counts %>% filter(category == "social & behavioral sciences") %>% 
  mutate(social_behavioral = prc_diversity) %>% right_join(bysbjrc_graph, by = "year")
bysbjrc_graph <- bysbjrc_prc_counts %>% filter(category == "other subjects") %>% 
  mutate(other_studies = prc_diversity) %>% right_join(bysbjrc_graph, by = "year")
# graphing percentages by collapsed cats 
figure_3B <- bysbjrc_graph %>% ggplot() + 
  geom_line(aes(y = genetics_heredity, x = year, 
                colour = "genetics & heredity"), size = 1.5, stat="identity") +
  geom_line(aes(y = biomedical_studies, x = year, 
                colour = "biomedical studies"), size = 1.5, stat="identity") +
  geom_line(aes(y = social_behavioral, x = year, 
                colour = "social & behavioral sciences"), size = 1.5, stat="identity") +
  geom_line(aes(y = other_studies, x = year, 
                colour = "other subjects"), size = 1.5, stat="identity") +
  theme_bw() +
  labs(title = "  Figure 3B. Proportional Growth of Population \n Terms by Collapsed Academic Field (1990-2017)",
       color =  "Field") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#E21F00", "#61B329", "#E2B306", "#3B9AB2")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_3AB <- bysbjrc_graph %>% ggplot() + 
  geom_line(aes(y = genetics_heredity, x = year, 
                colour = "genetics & heredity"), size = 1.5, stat="identity") +
  geom_line(aes(y = biomedical_studies, x = year, 
                colour = "biomedical studies"), size = 1.5, stat="identity") +
  geom_line(aes(y = social_behavioral, x = year, 
                colour = "social & behavioral sciences"), size = 1.5, stat="identity") +
  geom_line(aes(y = other_studies, x = year, 
                colour = "other subjects"), size = 1.5, stat="identity") +
  theme_bw() +
  labs(title = "     Figure 3B. Proportional Growth of Population \n Terms by Collapsed Academic Field (1990-2017)",
       color =  "Field") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "bottom",
        legend.spacing.x = unit(0.5, 'cm'),
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#E21F00", "#61B329", "#E2B306", "#3B9AB2")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_3AB <- get_legend(legend_3AB)

grid.arrange(figure_3A, figure_3B, 
             legend_3AB, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )

```

```{r figure 3cd, fig.height=5, fig.width=11.5, echo=FALSE}

# pulling our list of population
all_pop_terms <- read_csv("population_terms.csv") 
all_pop_terms <- paste(c("\\b(?i)(zcx", all_pop_terms$term, "zxc)\\b"), collapse = "|")

# standardizing all subjects  
pop_terms_bysubject <- abstract_data %>% 
  separate(subject, into = paste("subject", 1:15, sep = "_"), sep = ";") %>%
  gather(value, subject, subject_1:subject_15, na.rm = TRUE) %>% select(-value) %>% 
  separate(subject, into = c("subject", "void"), sep = "[(]") %>% select(-void) %>% 
  mutate(subject = stri_trim_both(subject)) %>% 
  mutate(subject = tolower(subject)) %>% 
  mutate(diversity = ifelse(test = str_detect(string = word, 
         pattern = "diversity"), yes = 1, no = 0)) %>% 
  mutate(term = ifelse(test = str_detect(string = word, 
         pattern = all_pop_terms), 
         yes = "all population terms", no = word)) %>%  
  mutate(all_pop_terms = ifelse(test = str_detect(string = term, 
         pattern = "\\b(all population terms)\\b"), yes = 1, no = 0))

# graphing 
figure_3C <- pop_terms_bysubject %>%   
  select(subject, term, year) %>%
  group_by(year) %>% 
  count(term, subject, sort = TRUE) %>% ungroup() %>%     
  filter(term == "all population terms") %>% 
  filter(subject == "genetics & heredity" | 
         subject == "biochemistry & molecular biology" | 
         subject == "immunology" |
         subject == "microbiology" | 
         subject == "pharmacology & pharmacy" | 
         subject == "infectious diseases" |  
         subject == "behavioral sciences" |
         subject == "sociology" ) %>% 
  ggplot() + geom_line(aes(y = n, x = year, colour = subject), size = 1.5, stat="identity") +
  theme_bw() +
  labs(title = "   Figure 3C. Proportional Growth of All \n Population Terms by Discipline (1990-2017)",
       color =  "Discipline") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Number of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#E2B306", "#E86F00", "#E21F00", "#6b0f6b", "#00468b", "#3B9AB2", "#d23ccc", "#61B329")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

# total articles each year 
bysubject_prc_counts <- text_data %>% group_by(year) %>% 
  count(year) %>% ungroup() %>% rename(total = n) 
# articles with term mentioned each year 
bysubject_prc_counts <- pop_terms_bysubject %>% filter(diversity == 1) %>% 
  group_by(subject, year) %>% summarise(cnt_diversity = n_distinct(pubmed_id)) %>% 
  right_join(bysubject_prc_counts, by = "year") %>% 
  mutate(prc_diversity = round(cnt_diversity / total * 100, digits = 2)) 
# wrangling all cats into a df 
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "genetics & heredity") %>% 
  mutate(genetics_heredity = prc_diversity) 
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "biochemistry & molecular biology") %>% 
  mutate(biochem_molbio = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "microbiology") %>% 
  mutate(microbio = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "infectious diseases") %>% 
  mutate(infectious = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "immunology") %>% 
  mutate(immunology = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "pharmacology & pharmacy") %>% 
  mutate(pharmacology = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "behavioral sciences") %>% 
  mutate(behavioral = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "health care sciences & services") %>% 
  mutate(healthcare = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "neurosciences & neurology") %>% 
  mutate(neurology = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "psychology") %>% 
  mutate(psychology = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "sociology") %>% 
  mutate(sociology = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "oncology") %>% 
  mutate(oncology = prc_diversity) %>% right_join(bysubject_graph, by = "year")
bysubject_graph <- bysubject_prc_counts %>% filter(subject == "business & economics") %>% 
  mutate(business = prc_diversity) %>% right_join(bysubject_graph, by = "year")

# graphing top-8 disciplines over time 
figure_3D <- bysubject_graph %>% 
  ggplot() + 
  geom_line(aes(y = genetics_heredity, x = year, 
                colour = "genetics & heredity"), size = 1.5, stat="identity") +
  geom_line(aes(y = biochem_molbio, x = year, 
                colour = "biochemistry & molecular biology"), size = 1.5, stat="identity") +
  geom_line(aes(y = microbio, x = year, 
                colour = "microbiology"), size = 1.5, stat="identity") +
  geom_line(aes(y = infectious, x = year, 
                colour = "infectious diseases"), size = 1.5, stat="identity") +
  geom_line(aes(y = immunology, x = year, 
                colour = "immunology"), size = 1.5, stat="identity") +
  geom_line(aes(y = pharmacology, x = year, 
                colour = "pharmacology & pharmacy"), size = 1.5, stat="identity") +
  geom_line(aes(y = behavioral, x = year, 
                colour = "behavioral sciences"), size = 1.5, stat="identity") +
  geom_line(aes(y = sociology, x = year, 
                colour = "sociology"), size = 1.5, stat="identity") +
  theme_bw() +
  labs(title = "   Figure 3D. Proportional Growth of All \n Population Terms by Discipline (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "none",
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#E2B306", "#E86F00", "#E21F00", "#6b0f6b", "#00468b", "#3B9AB2", "#d23ccc", "#61B329")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))


legend_3CD <- bysubject_graph %>% 
  ggplot() + 
  geom_line(aes(y = genetics_heredity, x = year, 
                colour = "genetics & heredity"), size = 1.5, stat="identity") +
  geom_line(aes(y = biochem_molbio, x = year, 
                colour = "biochemistry & molecular biology"), size = 1.5, stat="identity") +
  geom_line(aes(y = microbio, x = year, 
                colour = "microbiology"), size = 1.5, stat="identity") +
  geom_line(aes(y = infectious, x = year, 
                colour = "infectious diseases"), size = 1.5, stat="identity") +
  geom_line(aes(y = immunology, x = year, 
                colour = "immunology"), stat="identity") +
  geom_line(aes(y = pharmacology, x = year, 
                colour = "pharmacology & pharmacy"), size = 1.5, stat="identity") +
  geom_line(aes(y = behavioral, x = year, 
                colour = "behavioral sciences"), size = 1.5, stat="identity") +
  geom_line(aes(y = sociology, x = year, 
                colour = "sociology"), size = 1.5, stat="identity") +
  theme_bw() +
  labs(title = "   Figure 3D. Proportional Growth of All \n Population Terms by Discipline (1990-2017)",
       color =  "Term") + 
  #scale_linetype_manual( values=linetype) +
  ylab("Proportion of Articles") +
  theme(axis.title.x = element_blank(), 
        #axis.title.y = element_blank(), 
        legend.position = "bottom",
        legend.spacing.x = unit(0.5, 'cm'),
        legend.title = element_text(size = 10, hjust = 0.5, face="bold"),
        plot.title = element_text(size=12, hjust = 0.5, face="bold")) +
  scale_color_manual(values=c("#E2B306", "#E86F00", "#E21F00", "#6b0f6b", "#00468b", "#3B9AB2", "#d23ccc", "#61B329")) +
  scale_x_continuous(limits = c(1990, 2017),
                     breaks = c(1990, 1995, 2000, 2005, 2010, 2015))

legend_3CD <- get_legend(legend_3CD)

grid.arrange(figure_3C, figure_3D, 
             legend_3CD, 
             nrow = 2, ncol=2,
             layout_matrix = rbind(c(1,2), c(3,3)),
             #widths = c(2.7, 2.7), 
             heights = c(2.5, 0.4)
             )
#ggplotly(figure_3A)
```

```{r binding 3ab + 3cd, fig.height=7, fig.width=11.5, echo=FALSE}
lay <- rbind(c(1,2),
             c(3,3),
             c(4,5),
             c(6,6))

grid.arrange(figure_3A, figure_3B, 
             legend_3AB,
             figure_3C, figure_3D,
             legend_3CD,
             nrow = 4, ncol=2,
             layout_matrix = lay,
             #widths = c(2.7, 2.7), 
             heights = c(4.2, 0.4, 4.2, 0.8)
             )
```
```{r}


```

Supplementary Notes: Color Palettes

```{r supplementary color palettes, eval=FALSE}
test_pal <- c("#61B329", # first (green)
              "#3B9AB2", # second (Light Blue)
              "#E21F00", # third (Red )
              "#00468b", # fourth (Dark Blue)
              "#6b0f6b", # (purple)
              "#d23ccc", # pink 
              "#E86F00", # seventh (orange)
              "#E2B306") # last (yellow)

scale_color_manual(values=c("#6b0f6b", "#61B329", "#E21F00", "#E2B306", "#E86F00", "#3B9AB2",  "#00468b", "#d23ccc"))  # 8 color palette - 1A + 1B graphs 

scale_color_manual(values=c("#E21F00", "#3B9AB2", "#61B329", "#E2B306"))  # 4 color palette - 2A + 2B graphs

scale_color_manual(values=c("#61B329", "#3B9AB2", "#E2B306", "#E21F00", "#E86F00", "#00468b")) + # parsed diversity terms

scale_color_manual(values=c("#61B329", "#E2B306", "#3B9AB2", "#E86F00", "#E21F00")) + # international county analyses 

scale_color_manual(values=c("#61B329", "#3B9AB2", "#E21F00", "#00468b", "#d23ccc", "#6b0f6b",  "#E86F00", "#E2B306"))  # 8 color palette - us-specific graphs 
  
c("#E2B306", "#E86F00", "#E21F00", "#6b0f6b", "#00468b", "#3B9AB2", "#d23ccc", "#61B329")) + # discipline graphs 
  
c("#E21F00", "#E2B306", "#61B329", "#3B9AB2")) + #collapsed field 
  
```
